% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addECDFLinesToPlot.R
\name{addECDFLinesToPlot}
\alias{addECDFLinesToPlot}
\title{Add lines for the empirical cumulative density function to a plot}
\usage{
addECDFLinesToPlot(pn, lwParam)
}
\arguments{
\item{pn}{An empirical CDF of the type output by calcStoppingCDF}

\item{lwParam}{A parameter}
}
\description{
Point pairs for plotting an empirical cumulative density function (CDF; see \code{calcStoppingCDF} are created by \code{calcECDFPointPairs}. Add lines representing these point pairs to a plot.
}
\references{
\url{https://www.biorxiv.org/content/early/2017/01/31/104406}
}
\seealso{
calcStoppingCDF, calcCDFPointPairs
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
