% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayDem_loadCalibCurve.R
\name{bayDem_loadCalibCurve}
\alias{bayDem_loadCalibCurve}
\title{Load Calibration Curve}
\usage{
bayDem_loadCalibCurve(calibCurve)
}
\arguments{
\item{calibCurve}{Name of calibration curve}
}
\value{
The calibration dataframe, with columns yearBP, uncalYearBP, and uncalYearBPError
}
\description{
Parse and return the radiocarbon calibration curve stored in data
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
