% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{build_aic_output}
\alias{build_aic_output}
\title{Model selection using AIC}
\usage{
build_aic_output(
  data_dir,
  analysis_name,
  var_name,
  format_df = F,
  save_file = F
)
}
\arguments{
\item{data_dir}{Directory in which the model files are stored}

\item{analysis_name}{Unique ID for the given analysis}

\item{var_name}{Character string of the variable name for 
searching the directory}

\item{format_df}{Logical on whether the calculated AICs should be 
formatted as a dataframe. Default is format_df=FALSE, which maintains 
a list structure.}

\item{save_file}{Logical on whether the calculated AIC should be saved in
the directory as an .RDS file. Default is save_file=F.}
}
\value{
List or dataframe of the calculated AICs.
}
\description{
A wrapper function that imports all models given a variable, 
calculates the AIC for each model, and stores it in a list or dataframe.
}
