% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{build_model_vec}
\alias{build_model_vec}
\title{Build model vector}
\usage{
build_model_vec(mean_models, noise_models)
}
\arguments{
\item{mean_models}{A vector with the mean model names}

\item{noise_models}{A vector with the noise model names}
}
\value{
A vector with all possible mean-noise model combinations
}
\description{
A function to build a vector of mean-noise model specifications
to be used in the function [evaluate_univariate_models]. This function 
creates all possible mean-noise combinations, resulting in a vector of 
length = length(mean_models) x length(noise_models)
}
