% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{build_var_names}
\alias{build_var_names}
\title{Build a complete vector of variable names for the input data frame, var_info
(for details on var_info, see [load_var_info]}
\usage{
build_var_names(var_info)
}
\arguments{
\item{var_info}{The variable information data frame}
}
\value{
A list of all variable names specified by var_info, accounting for
  left/right variables
}
\description{
Build a complete vector of variable names for the input data frame, var_info
(for details on var_info, see [load_var_info]
}
\examples{
# Call build_var_names to generate expected column names in data
var_info_file <- syanalysis_name.file("extdata","US_var_info.csv",package="yada")
print(var_info_file)

var_info <- load_var_info(var_info_file)  # load file using file path
head(var_info)

var_names <- build_var_names(var_info)
var_names[5:16]

}
