% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{calc_aic}
\alias{calc_aic}
\title{Calculate AIC}
\usage{
calc_aic(k, eta)
}
\arguments{
\item{k}{Number of parameters in model}

\item{eta}{Negative log-likelihood}
}
\value{
AIC value
}
\description{
Function that calculates the Akaike's Information Criterion 
given the number of parameters (k) and negative log-likelihood (eta)
}
