% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{calc_bin_prob}
\alias{calc_bin_prob}
\title{Calculate the probability of the ordinal responses m = 0...M by binning
x-values and calculating the proportion of m-values in each bin.}
\usage{
calc_bin_prob(x, v, binBounds = NA)
}
\arguments{
\item{x}{The vector of independent variables}

\item{v}{The vector of responses (m-values)}

\item{binBounds}{(Optional) The bounds to use for binning. If not given, 20 evenly spaced bins on the interval min(x) to max(x) are used.}
}
\value{
A list with the bin centers, bin counts, and bin proportions
}
\description{
Calculate the probability of the ordinal responses m = 0...M by binning
x-values and calculating the proportion of m-values in each bin.
}
