% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{calc_ci_ord}
\alias{calc_ci_ord}
\title{Calculate ordinal model confidence intervals}
\usage{
calc_ci_ord(ord_model, th_x)
}
\arguments{
\item{ord_model}{List containing the parameter vector for y [th_y] and model
specification [mod_spec] for an ordinal univariate model (likely the 
output of `solve_ord_problem`)}

\item{th_x}{Parameterization for prior on x}
}
\description{
Wrapper function that takes an ordinal model and outputs the 
point estimate and 95% and 99% confidence intervals for each ordinal stage.
}
