% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_integral.R
\name{calc_conditional_gaussian_integral}
\alias{calc_conditional_gaussian_integral}
\title{Calculate the conditional multivariate Gaussian integral over a
rectangular domain}
\usage{
calc_conditional_gaussian_integral(
  mean_vect,
  cov_mat,
  lo = c(),
  hi = c(),
  y_giv = c(),
  log = FALSE
)
}
\arguments{
\item{mean_vect}{The vector of means [length J+K]}

\item{cov_mat}{The covariance matrix [dimension (J+K) by (J+K)]}

\item{lo}{The lower limit of integration for the dependent variables [length
J]}

\item{hi}{The upper limit of integration for the dependent variables [length
J]}

\item{y_giv}{The known values of the conditioned variables [length K]}

\item{log}{Whether to return the logarithm of the integral (default FALSE)}
}
\value{
The value of the integral
}
\description{
Calculate the multiviarate Gaussian integral over a rectangular domain where
some of the variables are known, and thus conditioned on. In particular,
there are two types of variables, dependent and given, and the integral is
over dependent variables. The known variables are provided as a vector ygiv,
with length K. The bounds of the integral over dependent variables are lo and
hi, each of which have length J. The multivariate Gaussian density is
specified by a mean vector, mean_vect, of length J+K, and a covariance
matrix, cov_mat, with dimensions (J+K) by (J+K). Edge case, such as J=0 and
K>0, for which which no integral is needed, are handled.
}
