% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{calc_eta_multivariate}
\alias{calc_eta_multivariate}
\title{Calculate multivariate (cindep and cdep) negative-log-likelihoods
for out-of-sample cross-validation or Akaike's Information Criterion.}
\usage{
calc_eta_multivariate(data_dir, analysis_name, fold = NA)
}
\arguments{
\item{data_dir}{Directory where the problem and model files are saved}

\item{analysis_name}{Unique identifier for current analysis}

\item{fold}{Fold number, if needed for cross-validation. (Default is NA 
for AIC)}
}
\description{
Function that loads the cindep and cdep models folds,  
performs negative-log-likelihood calculations, calculates AIC if defined, 
and returns either the maximum likelihood ("cv") or aic for each model.
}
