% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_joint.R
\name{calc_joint}
\alias{calc_joint}
\title{Calculate the joint probability for the mixed cumulative probit model}
\usage{
calc_joint(xcalc, y, theta_x, theta_y, hp)
}
\arguments{
\item{xcalc}{A vector of ages at which to calculate the posterior probability}

\item{y}{An observation of traits used for the posterior calculation}

\item{theta_x}{Parameterization for prior on x}

\item{theta_y}{Parameterization for likelihood}

\item{hp}{Hyperparameters}
}
\value{
A vector of joint probabilities
}
\description{
\code{calc_joint} calculates the joint density probability p(x,y|theta_x,theta_y) for the mixed cumulative probit model at the points in xcalc. xcalc is assumed to be evenly spaced.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
