% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{calc_kl_div}
\alias{calc_kl_div}
\title{Calculate the Kullback-Leibler (KL) divergence of a posterior density from a
prior density}
\usage{
calc_kl_div(x_post_obj, th_x)
}
\arguments{
\item{x_post_obj}{The result of a call to calc_x_posterior}

\item{th_w}{The parameterization of the prior}
}
\value{
The Kullback-Leibler (KL) divergence
}
\description{
The inputs are an output from \code{calc_x_posterior}, x_post_obj, and a
parameterization of the prior, th_x (see \code{calc_x_density}. The output
is the divergence of the posterior density from the prior density, which is
a measure of the information gained from the update step. The divergence is
calculated using a base 2 logarithm, so the units of the output are bits.
Any points for which the prior or posterior density is undefined,
non-finite, or zero are ignored.
}
