% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{calc_mean_univariate_cont}
\alias{calc_mean_univariate_cont}
\title{Calcuate the mean, h, for a univariate continuous model by extracting c from
th_w and calling calc_mean.}
\usage{
calc_mean_univariate_cont(x, th_w, modSpec)
}
\arguments{
\item{x}{The vector of independent variables}

\item{th_w}{The continuous parameter vector [c,kappa]}

\item{modSpec}{The model specification}
}
\value{
The mean, h(x,c), which is the same length as x
}
\description{
Calcuate the mean, h, for a univariate continuous model by extracting c from
th_w and calling calc_mean.
}
