% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mixed_cum_probit_posterior.R
\name{calc_mixed_cum_probit_posterior}
\alias{calc_mixed_cum_probit_posterior}
\title{Calculate posterior for mixed cumulative probit model}
\usage{
calc_mixed_cum_probit_posterior(theta_y, theta_x, xv, y, hp, asLog = F)
}
\arguments{
\item{theta_y}{theta_y as a list (single theta_y sample) or list of lists (multiple theta_y samples)}

\item{theta_x}{Parameterization for distribution of x (exponential rate)}

\item{xv}{A vector of ages at which to calculate posterior}

\item{y}{A new observation}

\item{hp}{Hyperparameters}

\item{asLog}{A logical variable indicating whether to return the (unnormalized) log of the likelihood}
}
\value{
The vector of posterior probabilities
}
\description{
\code{calc_mixed_cum_probit_posterior} calculates the posterior probability of an observation as a function of a vector xv, which is assumed to be uniformly sampled.
}
\seealso{
simConstHaz, fitConstHaz, fitLinHaz
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
