% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{calc_neg_log_lik_multivariate_core}
\alias{calc_neg_log_lik_multivariate_core}
\title{Calculate the vector of negative log-likelihoods values for a multivariate
model. To increase execution speed, the calculation data (calcData) output
by prep_for_neg_log_lik_multivariate is used.}
\usage{
calc_neg_log_lik_multivariate_core(th_y, x, y, mapping)
}
\arguments{
\item{th_y}{The parameter vector [a,tau,alpha,z]}

\item{x}{The independent variable}

\item{y}{The vector of responses}

\item{mapping}{The index mapping for rapid calculations}
}
\value{
The negative log-likelihood
}
\description{
Calculate the vector of negative log-likelihoods values for a multivariate
model. To increase execution speed, the calculation data (calcData) output
by prep_for_neg_log_lik_multivariate is used.
}
