% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{calc_neg_log_lik_ord}
\alias{calc_neg_log_lik_ord}
\title{Calcuate the negative log-likelihood}
\usage{
calc_neg_log_lik_ord(th_v, x, v, modSpec)
}
\arguments{
\item{th_v}{The ordinal parameter vector [b,tau,beta]}

\item{x}{The vector of independent variables}

\item{v}{The vector of ordinal responses}

\item{modSpec}{The model specification}
}
\value{
The negative log-likelihood
}
\description{
Calcuate the negative log-likelihood
}
