% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{calc_neg_log_lik_scalar_multivariate}
\alias{calc_neg_log_lik_scalar_multivariate}
\title{Calculate the negative log-likelihood for a single observation}
\usage{
calc_neg_log_lik_scalar_multivariate(th_y, calc_data_n)
}
\arguments{
\item{th_y}{The parameter vector}

\item{calc_data_n}{Data needed for a speedy negative log-likelihood
calculation (for a single observation, n)}
}
\value{
The negative log-likelihood for the observation
}
\description{
Calculate the negative log-likelihood for a single observation
}
