% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{calc_neg_log_lik_vect_cont}
\alias{calc_neg_log_lik_vect_cont}
\title{Calcuate the vector of negative log-likelihoods for a continuous model}
\usage{
calc_neg_log_lik_vect_cont(th_w, x, w, modSpec, tfCatVect = NA)
}
\arguments{
\item{x}{The vector of independent variables}

\item{w}{The vector of continuous responses}

\item{modSpec}{The model specification}

\item{tfCatVect}{(Optional) The transformation category vector to transform th_w to a constrained representation}

\item{th_v}{The continuous parameter vector [c,kappa]}
}
\value{
The vector of negative log-likelihoods, which is the same length as x
}
\description{
Calcuate the vector of negative log-likelihoods for a continuous model
}
