% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{calc_neg_log_lik_vect_multivariate_chunk_outer}
\alias{calc_neg_log_lik_vect_multivariate_chunk_outer}
\title{Use a parallel for loop along with "chunks" of input data to calculate the
vector of negative log-likelihoods. For each chunk, a conventional for loop
is used for each negative log-likelihood calculation and a parallel for loop
is used to iterate across chunks. The output is identical to
calc_neg_log_lik_vect_multivariate, but chunking the data can lead to better
performance on some systems.}
\usage{
calc_neg_log_lik_vect_multivariate_chunk_outer(
  th_y,
  calcData,
  tfCatVect = NA,
  numChunks = round(length(calcData)/10)
)
}
\arguments{
\item{th_y}{The parameter vector}

\item{calcData}{Data needed for a speedy negative log-likelihood calculation}

\item{tfCatVect}{A vector to transform th_y from an unconstrained to a constrained representation (if necessary)}

\item{nunChunks}{The number of chunks to use (default: round(length(calcData)/10))}
}
\value{
The vector of negative log-likelihoods
}
\description{
Use a parallel for loop along with "chunks" of input data to calculate the
vector of negative log-likelihoods. For each chunk, a conventional for loop
is used for each negative log-likelihood calculation and a parallel for loop
is used to iterate across chunks. The output is identical to
calc_neg_log_lik_vect_multivariate, but chunking the data can lead to better
performance on some systems.
}
