% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{calc_neg_log_lik_vect_ord}
\alias{calc_neg_log_lik_vect_ord}
\title{Calcuate the vector of negative log-likelihoods for an ordinal model}
\usage{
calc_neg_log_lik_vect_ord(th_v, x, v, modSpec, tfCatVect = NA)
}
\arguments{
\item{th_v}{The ordinal parameter vector [b,tau,beta]}

\item{x}{The vector of independent variables}

\item{v}{The vector of ordinal responses}

\item{modSpec}{The model specification}

\item{tfCatVect}{(Optional) The transformation category vector to transform th_v to a constrained representation}
}
\value{
The vector of negative log-likelihoods, which is the same length as x
}
\description{
Calcuate the vector of negative log-likelihoods for an ordinal model
}
