% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_new_x_log_density.R
\name{calc_new_x_log_density}
\alias{calc_new_x_log_density}
\title{Calculate the log density for a new observation x}
\usage{
calc_new_x_log_density(x, y, theta_y, theta_x, hp, transformVar = F)
}
\arguments{
\item{y}{The known value of the response variables}

\item{theta_y}{theta_y as a list (single theta_y sample) or list of lists (multiple theta_y samples)}

\item{theta_x}{Parameterization for distribution of x (exponential rate)}

\item{hp}{Hyperparameters}

\item{transformVar}{Boolean indicating whether the input is xbar [optional; default FALSE]}

\item{xbar}{log of the scalar age at which to calculate the log density}
}
\value{
The log density
}
\description{
\code{calc_new_x_log_density} calculates the log density for a new observation x given known y. This is in essence a wrapper for MCMC sampling of x for the probability density p(x|y,D,alpha). Optionally, the input can be the transformed variable xbar = log(x). The densities of x and x_bar are realated by p_xbar(xbar) = p_x(xbar)*exp(xbar).
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
