% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_functions.R
\name{calc_noise}
\alias{calc_noise}
\title{Calculate the noise, sigma, given x, the parameter vector alpha, and the
noise specification noiseSpec.}
\usage{
calc_noise(x, noiseSpec, alpha)
}
\arguments{
\item{x}{The independent variable}

\item{noiseSpec}{The noise specification}

\item{alpha}{The parameter vector}
}
\value{
sigma(x,alpha), the noise (same length as x)
}
\description{
Calculate the noise, sigma, given x, the parameter vector alpha, and the
noise specification noiseSpec.
}
