% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{calc_noise_univariate_ord}
\alias{calc_noise_univariate_ord}
\title{Calcuate the noise, gamma, for a univariate ordinal model by extracting beta
from th_v and calling calc_noise.}
\usage{
calc_noise_univariate_ord(x, th_v, modSpec)
}
\arguments{
\item{x}{The vector of independent variables}

\item{th_v}{The ordinal parameter vector [b,tau,beta]}

\item{modSpec}{The model specification}
}
\value{
The noise, gamma(x,beta), which is the same length as x
}
\description{
Calcuate the noise, gamma, for a univariate ordinal model by extracting beta
from th_v and calling calc_noise.
}
