% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{calc_q}
\alias{calc_q}
\title{Calculate the likelihood for a univariate ordinal model as a function of x}
\usage{
calc_q(x, th_v, m, mod_spec)
}
\arguments{
\item{x}{The vector of independent variables}

\item{th_v}{The ordinal parameter vector [b,tau,beta]}

\item{m}{The ordinal response value}

\item{mod_spec}{The model specification}
}
\value{
The vector for probabilities, which has the same length as x
}
\description{
Calculate the probability of measuring the response m as a function of x for
a univariate ordinal model
}
