% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_calculations.R
\name{calc_trapez_weights}
\alias{calc_trapez_weights}
\title{Calculate integration weights assuming the midpoint rule}
\usage{
calc_trapez_weights(x)
}
\arguments{
\item{x}{A vector of locations where the function is sampled, possibly
irregularly}
}
\value{
A vector of integration weights the same length as x
}
\description{
x is a vector of locations where a function to be integrated is evaluated
(the function values, f, are not input). Let x_n be the locations of
integration and f_n the corresponding function values for g=1, 2, ... N.
Assuming trapezoidal integration at the midpoints between elements of xu,
the weights to use for integration are dx_n = (x_(n+1) - x_(n-1)) / 2,
where the conventions x_0 = x_1 and x_(N+1) = tau_N are used.
}
