% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_calculations.R
\name{calc_weib_mix_density}
\alias{calc_weib_mix_density}
\title{Caculate the density for a Weibull mixture model}
\usage{
calc_weib_mix_density(x, z, theta)
}
\arguments{
\item{x}{Locations at which to evaluate probability density function}

\item{z}{Vector of mixture proportions}

\item{theta}{The value of shape and scale parameters with the ordering
(sh1_,sc_1,sh_2,sc_2,...)}
}
\value{
The PDF vector for the input vector x
}
\description{
Given the input vector x and Weibull mixture proportions and parameters z and
theta, calculate the probability density function for the input vector x.
}
