% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_calculations.R
\name{calc_weib_mix_density_matrix}
\alias{calc_weib_mix_density_matrix}
\title{Calculate the probability density function (PDF) matrix for a Weibull mixture
model}
\usage{
calc_weib_mix_density_matrix(x, theta)
}
\arguments{
\item{x}{Locations at which to evaluate probability density function}

\item{theta}{The value of shape and scale parameters with the ordering
(sh_1,sc_1,sh_2,sc_2,...)}
}
\value{
The PDF matrix with dimensions length(x) by length(theta)/2
}
\description{
Given the input vector x and Weibull mixture parameters theta, calculate the
probability density function (PDF) matrix, which has dimensions length(x) by
num_mix, where num_mix = length(theta)/2. The mixture proportions are not
accounted for in the calculation.
}
