% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{calc_x_density}
\alias{calc_x_density}
\title{Calculate the density at x given a parameterization for the density of
theta_x. Currently, the exponential distribution, Weibull mixtures, and
uniform distribution are supported.}
\usage{
calc_x_density(x, th_x)
}
\arguments{
\item{x}{A vector of ages at which to calculate the density}

\item{th_x}{A list with the fit type and parameter vector}
}
\description{
Calculate the density at x given a parameterization for the density of
theta_x. Currently, the exponential distribution, Weibull mixtures, and
uniform distribution are supported.
}
