% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{calc_x_posterior}
\alias{calc_x_posterior}
\title{Calculate the posterior probability p(x|y,theta_x,theta_y) for the mixed
cumulative probit model at the points in xcalc given the response vector
y. xcalc is assumed to be evenly spaced and the calculation is for a single
observation.}
\usage{
calc_x_posterior(xcalc, y, th_x, th_y, modSpec)
}
\arguments{
\item{xcalc}{A vector of ages at which to calculate the posterior probability}

\item{y}{The response vector for a single observation}

\item{th_x}{Parameterization for prior on x}

\item{th_y}{Parameterization for likelihood}

\item{modSpec}{The model specification}
}
\value{
A vector of posterior probabilities
}
\description{
Calculate the posterior probability p(x|y,theta_x,theta_y) for the mixed
cumulative probit model at the points in xcalc given the response vector
y. xcalc is assumed to be evenly spaced and the calculation is for a single
observation.
}
