% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{calc_x_prior}
\alias{calc_x_prior}
\title{Calculate the prior parameterization on x}
\usage{
calc_x_prior(
  data_dir,
  analysis_name,
  prior_type = "weibull",
  offset,
  seed,
  fold = NA,
  save_file = F
)
}
\arguments{
\item{data_dir}{The directory where data are pulled and saved}

\item{analysis_name}{Unique identifier for the analysis}

\item{prior_type}{The type of prior to be applied to x, currently only 
one available option, prior_type="weibull"}

\item{offset}{Amount by which to offset the x values}

\item{fold}{The fold number if not calculating for the main problem. 
Default is fold=NA}

\item{save_file}{Logical whether to save the output in data_dir}
}
\value{
A list with the prior parameterization specifications
}
\description{
A function to apply a Weibull offset mixture to the response 
  variable (x) to establish a parameterization for the prior on x
}
