% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCdfWeibMix.R
\name{calcCdfWeibMix}
\alias{calcCdfWeibMix}
\title{Calculate CDF given Weibull mixture proportions and parameters}
\usage{
calcCdfWeibMix(x, z, theta)
}
\arguments{
\item{x}{Locations at which to evaluate cumulative density function}

\item{z}{Vector of mixture proportions}

\item{theta}{The value of shape and scale parameters with the ordering [sh1,sc1,sh2,sc2,...]}
}
\value{
The CDF vector for the input vector x
}
\description{
Given the input vector x and Weibull mixture proportions and parameters z and theta, calculate the cumulative density function (CDF) for the input vector x.
}
\seealso{
\code{fitWeibMixMaxLikEM,calcPdfWeibMix,calcPdfMatrixWeibMix}
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
