% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcECDFPointPairs.R
\name{calcECDFPointPairs}
\alias{calcECDFPointPairs}
\title{Calculate empirical cumulative density function point pairs for plotting}
\usage{
calcECDFPointPairs(cdf)
}
\arguments{
\item{cdf}{An empirical CDF of the type output by calcStoppingCDF}
}
\value{
A matrix of point pairs with each row having (x1,y1,x2,y2) for a given point pair
}
\description{
Create point pairs for plotting the input empirical cumulative density function (CDF; see \code{calcStoppingCDF} for the form of the empirical CDF)
}
\references{
\url{https://www.biorxiv.org/content/early/2017/01/31/104406}
}
\seealso{
calcStoppingCDF, addECDFLinesToPlot
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
