% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLogLik_theta_y.R
\name{calcLogLik_theta_y}
\alias{calcLogLik_theta_y}
\title{Calculate the log-likelihood for theta_y in the mixed cumulative probit model}
\usage{
calcLogLik_theta_y(theta_y_list, x, Y, hp)
}
\arguments{
\item{theta_y_list}{theta_y as a list}

\item{x}{Vector of indepenent variable observations}

\item{Y}{Matrix of dependent variable observations}

\item{hp}{Hyperparameters}
}
\value{
The log-likelihood
}
\description{
\code{calcLogLik_theta_y} calculates the log-likelihood for theta_y (input as a list) given observed data.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
