% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMartResid.R
\name{calcMartResid}
\alias{calcMartResid}
\title{Calculate Martingale Residuals}
\usage{
calcMartResid(cumHaz, status)
}
\arguments{
\item{cumHaz}{A vector of cumulative hazards for the event of interest (length N)}

\item{status}{A vector of status indicators (length N)}
}
\value{
A vector of Martingale Residuals (length N)
}
\description{
Calculate Martingale Residuals for a vector of cumulative hazards and status indicators (0/FALSE for censoring and 1/TRUE for events)
}
\references{
\url{https://www.biorxiv.org/content/early/2017/01/31/104406}
}
\seealso{
calcMSR
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
