% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMSRDensityConstHaz.R
\name{calcMSRDensityConstHaz}
\alias{calcMSRDensityConstHaz}
\title{Calculate the Martingale Survival Residual probability density}
\usage{
calcMSRDensityConstHaz(D, rho, component = "both")
}
\arguments{
\item{D}{A vector of locations at which to calculate the probability density}

\item{rho}{The censoring ratio}

\item{component}{The component of the probability density to return. "both" yields the full density. "event" yields the component due to event occurrence. "censored" yields the component due to censoring. The default is "both".}
}
\value{
A vector of probability densities that is the same length as D
}
\description{
Calculate the expected probability density function for the Martingale Survival Residual (MSR). This assumes constant transition intensities (haard) for both event occurrence and censoring.
}
\references{
\url{https://www.biorxiv.org/content/early/2017/01/31/104406}
}
\seealso{
calcMSR
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
