% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNegLogLikWeibMix.R
\name{calcNegLogLikWeibMix}
\alias{calcNegLogLikWeibMix}
\title{Calculate the negative log-likelihood given Weibull mixture proportions and parameters}
\usage{
calcNegLogLikWeibMix(theta, x, z)
}
\arguments{
\item{theta}{The value of shape and scale parameters with the ordering [sh1,sc1,sh2,sc2,...]}

\item{x}{Locations at which to evaluate negative log-likelihood}

\item{z}{Vector of mixture proportions}
}
\value{
The negative log-likelihood (summed)
}
\description{
Given the input vector x and Weibull mixture proportions and parameters z and theta, calculate the negative log-likelihood.
}
\seealso{
\code{fitWeibMixMaxLikEM,calcPdfWeibMix}
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
