% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPdfMatrixWeibMix.R
\name{calcPdfMatrixWeibMix}
\alias{calcPdfMatrixWeibMix}
\title{Calculate PDF matrix given Weibull mixture parameters}
\usage{
calcPdfMatrixWeibMix(x, theta)
}
\arguments{
\item{x}{Locations at which to evaluate probability density function}

\item{theta}{The value of shape and scale parameters with the ordering [sh1,sc1,sh2,sc2,...]}
}
\value{
The PDF matrix with dimensions length(x) by length(theta)/2
}
\description{
Given the input vector x and Weibull mixture parameters theta, calculate the probability density function (PDF) matrix, which has dimensions length(x) by numMix, where numMix = length(theta)/2. The mixture proportions are not accounted for in the calculation.
}
\seealso{
\code{fitWeibMixMaxLikEM,calcPdfWeibMix}
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
