% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPdfWeibMix.R
\name{calcPdfWeibMix}
\alias{calcPdfWeibMix}
\title{Calculate PDF given Weibull mixture proportions and parameters}
\usage{
calcPdfWeibMix(x, z, theta)
}
\arguments{
\item{x}{Locations at which to evaluate probability density function}

\item{z}{Vector of mixture proportions}

\item{theta}{The value of shape and scale parameters with the ordering [sh1,sc1,sh2,sc2,...]}
}
\value{
The PDF vector for the input vector x
}
\description{
Given the input vector x and Weibull mixture proportions and parameters z and theta, calculate the probability density function for the input vector x.
}
\seealso{
\code{fitWeibMixMaxLikEM,calcPdfMatrixWeibMix}
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
