% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStoppingCDF.R
\name{calcStoppingCDF}
\alias{calcStoppingCDF}
\title{Calculate the empirical cumulative density function for a stopping distribution}
\usage{
calcStoppingCDF(y, halt, y.calc = NA, where = "after")
}
\arguments{
\item{y}{A vector of samples}

\item{halt}{A boolean vector indicating whether the stopping time was reached}

\item{y.calc}{Points at which to calculate the CDF}

\item{where}{Whether to calculate the CDF before or after the transition (or both)}
}
\value{
The CDF evaluated at the unique values of y (or y.calc) as a list with the following entries

\describe{
  \item{\code{y}}{The unique values of y or y.calc}
  \item{\code{before}}{The value of the CDF immediately before each y}
  \item{\code{after}}{The value of the CDF immediately after each y}
}
}
\description{
Calculate the empirical cumulative density function (CDF) of a sample of points (y) for use in the discontinuous Kolmogorov-Smirnov test at the points in y.calc. The samples are assumed to all be on the interval 0 to 1 (as for the Martingale Survival Residual (MSR)), and the CDF is assumed to be continuous up to a stopping time at which the transition is guaranteed to occur if it has not occurred already. Whether the stopping time was reached is indicated by the boolean vector halt. Duplicate values of y are acceptable. The optional input "where" indicates whether the CDF should be calculated after the  transition (default), instantaneously before the transition transition, or both; this only matters if y.calc contains values that are also in y (i.e, a calculation is done where a transition occurs in y). If y.calc is not specified, it is assumed to be all the values of y.
}
\references{
\url{https://www.biorxiv.org/content/early/2017/01/31/104406}
}
\seealso{
calcECDFPointPairs, addECDFLinesToPlot
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
