% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_y_transforms.R
\name{check_model}
\alias{check_model}
\title{Transformation functions for theta_y}
\usage{
check_model(paramModel)
}
\arguments{
\item{th_v}{Vector of parameters with ordering [rho,tau_1,...,tau_M,s,kappa]}
}
\description{
\code{powLawOrd}
}
\details{
Stuff
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
