% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R, R/data_io.R
\name{clear_temp_dir}
\alias{clear_temp_dir}
\title{Clear all the files in the temporary directory}
\usage{
clear_temp_dir()

clear_temp_dir()
}
\description{
Clear all the files in the temporary directory. Directories are not deleted.
clear_temp_dir is used in testing.

Clear all the files in the temporary directory. Directories are not deleted.
clear_temp_dir is used in testing.
}
