% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrVect2CovMat.R
\name{corrVect2CovMat}
\alias{corrVect2CovMat}
\title{Convert from correlation vector to covariance matrix}
\usage{
corrVect2CovMat(s)
}
\arguments{
\item{s}{The correlation vector}

\item{Sigma}{The covariance matrix}
}
\value{
The covariance matrix Sigma
}
\description{
\code{corrVect2CovMat} converts the input correlation vector into an equivalent covariance matrix Sigma. The diagonal terms of Sigma are $1$. The ordering is by row: Sigma_12, Sigma_13,...,Sigma_23,....
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
