% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMat2CorrVect.R
\name{covMat2CorrVect}
\alias{covMat2CorrVect}
\title{Convert from covariance matrix to correlation vector}
\usage{
covMat2CorrVect(Sigma)
}
\arguments{
\item{Sigma}{The covariance matrix}

\item{s}{The correlation vector}
}
\value{
The correlation vector s
}
\description{
\code{covMat2corrVect} converts the input covariance matrix Sigma into an equivalent vector of correlations (off diagonal terms). The diagonal terms should be $1$. The ordering is by row: Sigma_12, Sigma_13,...,Sigma_23,....
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
