% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{crossval_multivariate_models}
\alias{crossval_multivariate_models}
\title{Calculate multivariate (cindep and cdep) negative-log-likelihoods
for cross-validation}
\usage{
crossval_multivariate_models(data_dir, analysis_name, fold)
}
\arguments{
\item{data_dir}{Directory where the problem and model files are saved}

\item{analysis_name}{Unique identifier for current analysis}

\item{fold}{Fold number}
}
\description{
Function that loads the cindep and cdep models folds,  
performs negative-log-likelihood calculations, and returns vectors for
each corresponding model
}
