% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{doGpBasedOptim}
\alias{doGpBasedOptim}
\title{Solve an unconstrained minimization problem that involves random selection of
candidates followed by assessment of candidates using a Gaussian Process (GP)
model based on previous calculations of the objective function (to increase
estimation and evaluation speed of the GP, the fast local approximation of
laGP package is used). Although  it is assumed that the optimization problem
is unconstrained, the objective function can return Inf or NA (such points
are not used in the GP model). The behavior of the function is controlled
by the variable control, which is a list with named variables. conrol$Nmodel
specifies the number of preceding observations to use for the GP model. 
control$Ncand specifies the number of candidates to check, where candidates
are randomly chosen with a normal draw for which the mean equals the current
best parameter vector and the standard deviations are specified by
control$scaleVect. For rerpoducibility, the control$seed can be used to set
the random number seed. The candidate with the lowest estimated value of the
objective function is chosen for the next evaluation of the objective
function. The problem can be initialized in one of two ways: by specifying
an initial parameter vector, param0, or from a save file storing the results
of a previous "leg" of the optimization. If an initial parameter vector is
used for initialization, the objective function must be specified using the
optional input objFun. If a save file is used, the objective function does
not need to be input (if it is, it is ignored and a warning is given). The
optimization terminates either after the maximum number of evaulations,
control$maxEval, have been made, or the objective function has failed to
improve more than the objective function tolerance, control$objTol, over the
last 4*control$Nmodel evaluations. The boolean variable control$catchErrors
allows catching errors in the call to objFun using try, and setting the
ojective function value to Inf (infinity) if an error is encountered.
control$verbose specifies whether or not to print out summary information as
the optimization proceeds. control$showPlot specifies whether to plot the
progress of the optimization algorithm as it proceeds. The optional input
saveFile specifies whether to save the state of the optimization after each
function evaluation (default NA, not used). The control varibles and their
default values are:
****************************************************************************
scaleVect   Scaling vector of standard deviations for drawing new candidate
            parameter vectors
Default:    rep(1e-4,length(param0)
****************************************************************************
Nmodel      Number of observations to use to build the Gaussian Process model
  Default:    100
****************************************************************************
Ncand       Number of randomly selected candidates to check with the Gaussian
            Process model 
  Default:    100
*****************************************************************************
maxEval     Maximum number of evaluations of the objective function
  Default:    10000
*****************************************************************************
objTol      Tolerance to use in the stopping condition
  Default:    1e-4
*****************************************************************************
catchErrors Whether or not to catch errors in the objective function
            evaluation, and set the value to Inf (infinity) if an error
            is encountered
  Default:    FALSE
*****************************************************************************
seed        The random number seed to use. If it is not input, a random
            integer on the interval 1 to 1000000 is used (seeds from previous
            are never reused)
  Default:    sample.int(1000000,1)
*****************************************************************************
verbose     Whether or not to print out progress of the optimization
  Default:    FALSE
*****************************************************************************
showPlot    Whether or not to show a plot of the objective function value as
            the optimization proceeds
  Default:    FALSE
*****************************************************************************
The named variables in the output list are:
*****************************************************************************
paramMat    A matrix of parameter values for the last control$Nmodel
            evaluations that did not evaluate to infinity (or as many are
            available if fewer than control$Nmodel)
*****************************************************************************
objVect     A vector of objective function values for the last control$Nmodel
            evaluations that did not evaluate to infinity (or as many are
            available if fewer than control$Nmodel)
*****************************************************************************
objVectFull A vector of all the objective function values. If an evaluation
            is infinite, NA is stored.
*****************************************************************************
param_best  The parameter vector of the best solution found so far
*****************************************************************************
f_best      The objective function value of the best solution found so far
*****************************************************************************
legData     Information specific to each leg (call) to the doGpBasedOptim.
            legData is a list with a length equal to the number of legs. For
            all legs its elements  include control0, the input control
            variable, control, the full list of used control variables, and
            saveFile, the save file to use (NA if none is used). In addition,
            for the first leg the variables param0, the initial parameter
            vector, f0, the value of the objective function at param0, and
            objFun, the bjective function, are also stored.
*****************************************************************************}
\usage{
doGpBasedOptim(init, objFun = NA, control = list(), saveFile = NA, ...)
}
\arguments{
\item{init}{The initialization object (either an initial parameter vector or a save file from a previous leg of the optimization)}

\item{objFun}{The objective function (ignored if init is a save file)}

\item{control}{A named list of control variables (see function description for details)}

\item{saveFile}{An optional save file to store intermediate results of the optimization (same variables as output and the default is NA, no save file used)}

\item{...}{Additional inputs to the objective function}
}
\value{
A list of named variables containing key information about the optimization (see details)
}
\description{
Solve an unconstrained minimization problem that involves random selection of
candidates followed by assessment of candidates using a Gaussian Process (GP)
model based on previous calculations of the objective function (to increase
estimation and evaluation speed of the GP, the fast local approximation of
laGP package is used). Although  it is assumed that the optimization problem
is unconstrained, the objective function can return Inf or NA (such points
are not used in the GP model). The behavior of the function is controlled
by the variable control, which is a list with named variables. conrol$Nmodel
specifies the number of preceding observations to use for the GP model. 
control$Ncand specifies the number of candidates to check, where candidates
are randomly chosen with a normal draw for which the mean equals the current
best parameter vector and the standard deviations are specified by
control$scaleVect. For rerpoducibility, the control$seed can be used to set
the random number seed. The candidate with the lowest estimated value of the
objective function is chosen for the next evaluation of the objective
function. The problem can be initialized in one of two ways: by specifying
an initial parameter vector, param0, or from a save file storing the results
of a previous "leg" of the optimization. If an initial parameter vector is
used for initialization, the objective function must be specified using the
optional input objFun. If a save file is used, the objective function does
not need to be input (if it is, it is ignored and a warning is given). The
optimization terminates either after the maximum number of evaulations,
control$maxEval, have been made, or the objective function has failed to
improve more than the objective function tolerance, control$objTol, over the
last 4*control$Nmodel evaluations. The boolean variable control$catchErrors
allows catching errors in the call to objFun using try, and setting the
ojective function value to Inf (infinity) if an error is encountered.
control$verbose specifies whether or not to print out summary information as
the optimization proceeds. control$showPlot specifies whether to plot the
progress of the optimization algorithm as it proceeds. The optional input
saveFile specifies whether to save the state of the optimization after each
function evaluation (default NA, not used). The control varibles and their
default values are:
****************************************************************************
scaleVect   Scaling vector of standard deviations for drawing new candidate
            parameter vectors
Default:    rep(1e-4,length(param0)
****************************************************************************
Nmodel      Number of observations to use to build the Gaussian Process model
  Default:    100
****************************************************************************
Ncand       Number of randomly selected candidates to check with the Gaussian
            Process model 
  Default:    100
*****************************************************************************
maxEval     Maximum number of evaluations of the objective function
  Default:    10000
*****************************************************************************
objTol      Tolerance to use in the stopping condition
  Default:    1e-4
*****************************************************************************
catchErrors Whether or not to catch errors in the objective function
            evaluation, and set the value to Inf (infinity) if an error
            is encountered
  Default:    FALSE
*****************************************************************************
seed        The random number seed to use. If it is not input, a random
            integer on the interval 1 to 1000000 is used (seeds from previous
            are never reused)
  Default:    sample.int(1000000,1)
*****************************************************************************
verbose     Whether or not to print out progress of the optimization
  Default:    FALSE
*****************************************************************************
showPlot    Whether or not to show a plot of the objective function value as
            the optimization proceeds
  Default:    FALSE
*****************************************************************************
The named variables in the output list are:
*****************************************************************************
paramMat    A matrix of parameter values for the last control$Nmodel
            evaluations that did not evaluate to infinity (or as many are
            available if fewer than control$Nmodel)
*****************************************************************************
objVect     A vector of objective function values for the last control$Nmodel
            evaluations that did not evaluate to infinity (or as many are
            available if fewer than control$Nmodel)
*****************************************************************************
objVectFull A vector of all the objective function values. If an evaluation
            is infinite, NA is stored.
*****************************************************************************
param_best  The parameter vector of the best solution found so far
*****************************************************************************
f_best      The objective function value of the best solution found so far
*****************************************************************************
legData     Information specific to each leg (call) to the doGpBasedOptim.
            legData is a list with a length equal to the number of legs. For
            all legs its elements  include control0, the input control
            variable, control, the full list of used control variables, and
            saveFile, the save file to use (NA if none is used). In addition,
            for the first leg the variables param0, the initial parameter
            vector, f0, the value of the objective function at param0, and
            objFun, the bjective function, are also stored.
*****************************************************************************
}
