% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinadics.R
\name{dual}
\alias{dual}
\title{Calculate the dual}
\usage{
dual(m, num_elements = NA, N = NA, k = NA)
}
\arguments{
\item{m}{The lexical index of each unique choice (starting from 0)}

\item{num_elements}{The number of things that can be chosen (default: NA, not
used)}

\item{N}{The number of unique things (default: NA, not used)}

\item{k}{The number of things to choose out of N total things (default: NA,
  not used)

return The dual, d (see details for [yada::comb]}

\item{d}{The dual(see details)}

\item{e}{The element (a vector; see details)}
}
\description{
For an overview of combinadics and the dual see [yada::comb]. To calculate
the dual, the number of elements that can be chosen, num_elements, must be
known. It is either input directly or determined from N and k, which are
input. In total, therefore, there are two valid inputs patterns for the three
optional inputs num_elements, N, and k: specify numElements or specify both N
and k. If all three optional inputs are given, N and k are ignored.
}
