% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinadics.R
\name{elem}
\alias{elem}
\title{Calculate the element}
\usage{
elem(m, N, k)
}
\arguments{
\item{m}{The lexical index of each unique choice (starting from 0)}

\item{N}{The number of unique things}

\item{k}{The number of things to choose out of N total things

return The element, e (a vector; see details for [yada::comb]}
}
\description{
For an overview of combinadics and the element see [yada::comb].
}
