% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinadics.R
\name{elem_to_index}
\alias{elem_to_index}
\title{Convert the element (a vector) to the corresponding index (a scalar integer)}
\usage{
elem_to_index(e, N)
}
\arguments{
\item{N}{The number of unique things

return The index, a scalar integer (see details for [yada::comb]}

\item{m}{The element (a vector)}
}
\description{
For an overview of combinadics and the element/index see [yada::comb].
}
