% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_theta_y.R
\name{estimate_theta_y}
\alias{estimate_theta_y}
\title{Sample theta_y in the mixed cumulative probit model}
\usage{
estimate_theta_y(x, Y, hp, verbose = T)
}
\arguments{
\item{x}{Vector of indepenent variable observations}

\item{Y}{Matrix of dependent variable observations}

\item{hp}{Hyperparameters}

\item{verbose}{Whether to print out information as the run proceeds [optional]}

\item{numSamp}{Number of samples}

\item{burnIn}{Number of burn-in samples}

\item{start}{Starting value of theta_y_list for sampling [optional]}

\item{fileName}{File name for saving [optional]}

\item{savePeriod}{Period (of samples) for saving [optiona; default 1000]}

\item{known}{Known values, a list with the field theta_y, which is also a list [optional]}

\item{varNames}{Variable names for saving to file [optional]}
}
\value{
theta_y as a list
}
\description{
\code{sample_theta_y} uses the adaptive Metropolis algorithm of Haario et al. 2001, https://projecteuclid.org/euclid.bj/1080222083, to sample the vector theta_y for the cumulative probit model.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
