% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{evaluate_multivariate_models}
\alias{evaluate_multivariate_models}
\title{Evaluation of multivariate models}
\usage{
evaluate_multivariate_models(data_dir, analysis_name, eval_type = "aic")
}
\arguments{
\item{data_dir}{Directory where the problem and model files are saved}

\item{analysis_name}{Unique identifier for current analysis}

\item{eval_type}{Whether model selection should use out-of-sample 
cross-validation ("cv") or Akaike's Information Criterion ("aic")}
}
\description{
[evaluate_multivariate_models] utilizes either out-of-sample
cross-validation ("cv") or Akaike's Information Criterion ("aic") to 
aid in multivariate model selection between the conditionally independent 
("cindep") and conditionally dependent ("cdep") models generated by yada.
}
