% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_transforms.R
\name{extract_univariate_param_vect}
\alias{extract_univariate_param_vect}
\title{Extract the parameter vector for a univariate fit from a larger parameter
vector}
\usage{
extract_univariate_param_vect(th_y, mod_spec, j = NA, k = NA)
}
\arguments{
\item{th_y}{The full parameter vector}

\item{mod_spec}{The full model specification}

\item{j}{The index of the ordinal variable to extract (default: NA)}

\item{k}{The index of the ordinal variable to extract (default: NA)}
}
\value{
The univariate parameter vector
}
\description{
Extract the parameter vector for a univariate fit from a larger parameter
vector
}
