% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{fail_for_beta2}
\alias{fail_for_beta2}
\title{A helper function to check whether the fits in TH_y fail for beta2}
\usage{
fail_for_beta2(TH_y, beta2_max)
}
\arguments{
\item{TH_y}{TH_y is a matrix with dimensions num_param x (1+num_folds)}

\item{beta2_max}{beta2 threshold}
}
\description{
A helper function to check whether the fits in TH_y fail for beta2
}
