% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{fit_multivariate}
\alias{fit_multivariate}
\title{Fit a multivariate mixed cumulative probit model.}
\usage{
fit_multivariate(x, Y, modSpec, optimControl = list())
}
\arguments{
\item{x}{The vector of independent variables}

\item{Y}{The matrix of responses}

\item{modSpec}{The model specification}

\item{optimControl}{Control information for the call to doGpBasedOptim}
}
\value{
A list object of simulated data containing x, Y, and Ystar
}
\description{
Fit a multivariate mixed cumulative probit model.
}
