% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{fit_univariate_cont}
\alias{fit_univariate_cont}
\title{Fit a continuous model}
\usage{
fit_univariate_cont(x, w, modSpec, reqConv = T)
}
\arguments{
\item{x}{The vector of independent variables}

\item{w}{The vector of continuous responses}

\item{modSpec}{The model specification}

\item{reqConv}{Whether to require convergence of the optimization}
}
\value{
The continuous parameter vector, th_w
}
\description{
Fit a continuous model
}
