% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{fit_univariate_ord}
\alias{fit_univariate_ord}
\title{Fit a univariate ordinal model}
\usage{
fit_univariate_ord(x, v, modSpec, reqConv = T)
}
\arguments{
\item{x}{The vector of independent variables}

\item{v}{The vector of ordinal responses}

\item{modSpec}{The model specification}

\item{reqConv}{Whether to require convergence of the optimization}
}
\value{
The ordinal parameter vector, th_v
}
\description{
Fit a univariate ordinal model
}
