% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitConstHaz.R
\name{fitConstHaz}
\alias{fitConstHaz}
\title{Fit event history data assuming constant event and censoring hazards}
\usage{
fitConstHaz(survObj)
}
\arguments{
\item{survObj}{A Surv object (survival package) with event times and a status indicator (see \code{simConstHaz})}
}
\value{
A list with three fields

\describe{
  \item{\code{b}}{Overall, combined event and censoring rate}
  \item{\code{b1}}{Event rate}
  \item{\code{b2}}{Censoring rate}
}
}
\description{
\code{fitConstHaz} does a maximum likelihood fit of the input data assuming constant event and censoring hazards. The input is a Surv object (package survival) with a time column and a status column. Time is the event or censoring time. Status is 1 (or TRUE) if the event occured and 0 (or FALSE) if censoring occurred.
}
\seealso{
simConstHaz
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
