% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSilerMaxLik.R
\name{fitSilerMaxLik}
\alias{fitSilerMaxLik}
\title{Maximum likelihood fit for Siler distribution}
\usage{
fitSilerMaxLik(x, a0 = c(0.175, 1.4, 0.368 * 0.01, 0.075 * 0.001, 0.917 *
  0.1), x0 = 0, calcHessian = FALSE)
}
\arguments{
\item{x}{Locations at which to evaluate probability density function}

\item{a0}{Initial value for optimization. Default from Gage and Dyke 1986, Table 2, Level 15}

\item{x0}{x-offset. The default is 0 (no offset).}

\item{calcHessian}{Whether to calculate the Hessian}
}
\value{
A list consisting of the fit (on the transformed variable abar) and maximum likelihood estimate of a. Optionally, the Hessian of the log-likelihood is returned, which allows estimation of the standard errors of the maximum likelihood estimate via the observed Fisher information matrix.
}
\description{
Do a maximum likelihood Siler fit for the input data vector x. To allow unconstrained optimization to be used, optimization is done on the transformed variable abar = log(a/a0).
}
\seealso{
\code{dsiler}
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
