% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitWeibMixMaxLikEM.R
\name{fitWeibMixMaxLikEM}
\alias{fitWeibMixMaxLikEM}
\title{Maximum likelihood fit for Weibull mixture}
\usage{
fitWeibMixMaxLikEM(x, z0, theta0, numSteps = 25)
}
\arguments{
\item{x}{Locations at which to evaluate probability density function}

\item{z0}{Initial value of mixture proportions}

\item{theta0}{The value of shape and scale parameters with the ordering [sh1,sc1,sh2,sc2,...]}

\item{numSteps}{Number of EM steps to perform (default 25)}
}
\value{
A list consisting of the final z (mixture proportions), theta (mixture parameters), and nll (negative log-likelihood).
}
\description{
Do a maximum likelihood Weibull mixture fit for the input data vector x using the Expectation-Maximization (EM) alogrithm.
}
\seealso{
\code{calcPdfMatrixWeibMix,calcPdfWeibMix,calcNegLogLikWeibMix}
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
