% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCrra.R
\name{genCrra}
\alias{genCrra}
\title{Generalized CRRA parametric model}
\usage{
genCrra(x, param, transformW = F)
}
\arguments{
\item{x}{Vector of independent variable observations}

\item{y}{Vector of dependent variable observations}

\item{a}{Multiplicative coefficient}

\item{r}{Scaling exponent}

\item{b}{Offset}
}
\description{
\code{genCrra} calculates the response (y). fitGenCrra returns the least squares fit. \code{simGenCrra} creates simulated data.
}
\details{
The parametric form of the generalized Coefficient of Relative Risk Aversion (CRRA) is:

\deqn{y = a*x^(1-r) + b}

where \deqn{x} is the independent variable, \deqn{y} the dependent (response) variable, \deqn{a} the multiplicative coefficient, \deqn{r} the scaling exponent, and \deqn{b} the offset.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
