% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{generate_cv_problems}
\alias{generate_cv_problems}
\title{A wrapper to generate the cross-validation problems}
\usage{
generate_cv_problems(main_problem, K, seed = NA)
}
\arguments{
\item{main_problem}{The main problem}

\item{K}{Number of cross-validation folds}

\item{seed}{Setting a seed to allow for reproducibility, which is stored
within each problem fold. NA generates a randomly-selected seed between 1 and 1E6.}
}
\value{
A list with test and training data sets
}
\description{
A wrapper to generate the cross-validation problems
}
