% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{generate_main_problem}
\alias{generate_main_problem}
\title{Generate the main problem file}
\usage{
generate_main_problem(data_file, var_info)
}
\arguments{
\item{data_file}{Data frame containing the training data}

\item{var_info}{The var_info file loaded via [load_var_info()]}
}
\value{
A list containing the response variable (x) as a vector, 
  the predictor variables as a matrix (Y), and another list (mod_spec) 
  containing the model specifications.
}
\description{
The main problem file is a list containing the training data 
  for the mcp algorithm, as well as model specifications ("mod_spec") 
  including the number of ordinal (J) and continuous (K) variables and 
  the number of breaks between ordinal stages (M).
}
