% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{generate_mod_spec}
\alias{generate_mod_spec}
\title{Generate mod_spec list for use in yada functions}
\usage{
generate_mod_spec(problem, model_params, cdep_spec, uni_variable = NA)
}
\arguments{
\item{problem}{Object containing the problem file for current mod_spec}

\item{model_params}{Data frame containing variable parameters for the best
performing models (generated using `get_best_univariate_params()`)}

\item{cdep_spec}{Distinction of a conditionally independent ('indep') or
conditionally dependent ('dep') model.}

\item{uni_variable}{Trait name used for univariate model. If left as 'NA', the
mod_spec for multivariate models is generated.}
}
\description{
A simple function that provides assistance in generating the
list-format for mod_spec expected by yada functions
}
