% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{generate_ord_ci}
\alias{generate_ord_ci}
\title{Generate the confidence intervals for a given a cross-validated
ordinal variable}
\usage{
generate_ord_ci(data_dir, analysis_name, var_name, th_x, save_file = F)
}
\arguments{
\item{data_dir}{Directory where files are saved}

\item{analysis_name}{Unique identifier for current analysis}

\item{var_name}{Response variable name}

\item{th_x}{Parameterization for prior on x}

\item{save_file}{Logical whether the resulting data frame should be saved
as an .rds file}
}
\description{
This function returns the point estimate, 95% and 99%
confidence intervals for each ordinal stage of a given response variable.
}
