% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{get_best_univariate_params}
\alias{get_best_univariate_params}
\title{Get the parameters of the best cross-validated fits}
\usage{
get_best_univariate_params(data_dir, analysis_name, save_file = TRUE)
}
\arguments{
\item{data_dir}{Directory to save univariate model parameters}

\item{analysis_name}{analysis_name for file-naming}

\item{save_file}{Logical whether to save file or not as .rds}
}
\value{
Data frame with variable (var), type, mean_spec, noise_spec, parameter
  name (params), and parameter value (param_val).
}
\description{
Wrapper function that takes object cv_data, stores the highest ranking model
specification, noise specification, and model variables, and outputs all
results as a RDS file in dataframe format.
}
