% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{get_J}
\alias{get_J}
\title{Get the number of ordinal variables given a model specification. If J is not
a field in modSpec, 0 is returned.}
\usage{
get_J(modSpec)
}
\arguments{
\item{modSpec}{The model specification}
}
\value{
The number of ordinal variables, J
}
\description{
Get the number of ordinal variables given a model specification. If J is not
a field in modSpec, 0 is returned.
}
