% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{get_K}
\alias{get_K}
\title{Get the number of continuous variables given a model specification. If K is
not a field in modSpec, 0 is returned.}
\usage{
get_K(modSpec)
}
\arguments{
\item{modSpec}{The model specification}
}
\value{
The number of continuous variables, K
}
\description{
Get the number of continuous variables given a model specification. If K is
not a field in modSpec, 0 is returned.
}
