% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_functions.R
\name{get_mean_transform_categories}
\alias{get_mean_transform_categories}
\title{For a given mean specification, return a vector that gives the categories
used by param_constr2uncontr and param_constr2unconstr.}
\usage{
get_mean_transform_categories(meanSpec)
}
\arguments{
\item{meanSpec}{The mean specification}
}
\value{
A vector of categories
}
\description{
For a given mean specification, return a vector that gives the categories
used by param_constr2uncontr and param_constr2unconstr.
}
