% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_functions.R
\name{get_mean_transform_matrix}
\alias{get_mean_transform_matrix}
\title{Get a numeric matrix that specifies the transforms as described in
param_constr2uncontr and param_constr2unconstr.}
\usage{
get_mean_transform_matrix(meanSpec)
}
\arguments{
\item{meanSpec}{The mean specification}
}
\value{
A matrix specifying the transform
}
\description{
Get a numeric matrix that specifies the transforms as described in
param_constr2uncontr and param_constr2unconstr.
}
