% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{get_multivariate_transform_categories}
\alias{get_multivariate_transform_categories}
\title{For a given model specification of a multivariate model, return a vector that
gives the categories used by param_constr2uncontr and param_constr2unconstr.}
\usage{
get_multivariate_transform_categories(modSpec)
}
\arguments{
\item{modSpec}{The model specification}
}
\value{
A vector of categories
}
\description{
For a given model specification of a multivariate model, return a vector that
gives the categories used by param_constr2uncontr and param_constr2unconstr.
}
