% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_functions.R
\name{get_noise_transform_categories}
\alias{get_noise_transform_categories}
\title{For a given noise specification, return a vector that gives the categories
used by param_constr2uncontr and param_constr2unconstr.}
\usage{
get_noise_transform_categories(noiseSpec)
}
\arguments{
\item{noiseSpec}{The noise specification}
}
\value{
A vector of categories
}
\description{
For a given noise specification, return a vector that gives the categories
used by param_constr2uncontr and param_constr2unconstr.
}
