% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_functions.R
\name{get_noise_transform_matrix}
\alias{get_noise_transform_matrix}
\title{Get a numeric matrix that specifies the transforms as described in
param_constr2uncontr and param_constr2unconstr.}
\usage{
get_noise_transform_matrix(noiseSpec)
}
\arguments{
\item{noiseSpec}{The noise specification}
}
\value{
A matrix specifying the transform
}
\description{
Get a numeric matrix that specifies the transforms as described in
param_constr2uncontr and param_constr2unconstr.
}
