% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{get_num_folds}
\alias{get_num_folds}
\title{A helper function to identify univariate fits in a folder and determine the
number of folds for each fit.}
\usage{
get_num_folds(data_dir, analysis_name, j = NA, k = NA)
}
\arguments{
\item{data_dir}{Directory where the files are located.}

\item{analysis_name}{analysis_name for file-searching}

\item{j}{number of current ordinal variable}

\item{k}{number of current continuous variable}
}
\description{
A helper function to identify univariate fits in a folder and determine the
number of folds for each fit.
}
