% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{get_num_training_problems}
\alias{get_num_training_problems}
\title{A helper function to identify the number of training problems (folds)}
\usage{
get_num_training_problems(data_dir, analysis_name)
}
\arguments{
\item{data_dir}{The data directory with problems and results}

\item{analysis_name}{A unique analysis name (for the input data directory)}
}
\value{
Number of test problems (folds)
}
\description{
The training problem files begin with "train_analysis_name_fold". Count the
number of such files.
}
