% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_functions.R
\name{get_num_var_mean}
\alias{get_num_var_mean}
\title{Return the number of parameters for the mean specification meanSpec.}
\usage{
get_num_var_mean(meanSpec)
}
\arguments{
\item{meanSpec}{The mean specification}
}
\value{
The number of mean parameters
}
\description{
Return the number of parameters for the mean specification meanSpec.
}
