% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{get_num_var_multivariate}
\alias{get_num_var_multivariate}
\title{Get the number of parameters for named variables of a multivariate model
model}
\usage{
get_num_var_multivariate(
  varName,
  modSpec,
  j = NA,
  k = NA,
  i = NA,
  preceding = F
)
}
\arguments{
\item{varName}{The variable name (a, tau, alpha, z)}

\item{modSpec}{The model specification}

\item{j}{The ordinal index}

\item{k}{The continuous index}

\item{i}{The overall index}

\item{preceding}{Whether to return the number of precceding variables}
}
\value{
The number of parameters
}
\description{
Get the number of parameters for named variables of a multivariate model
model
}
