% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_functions.R
\name{get_num_var_noise}
\alias{get_num_var_noise}
\title{Return the number of parameters for the noise specification noiseSpec.}
\usage{
get_num_var_noise(noiseSpec)
}
\arguments{
\item{noiseSpec}{The noise specification}
}
\value{
The number of noise parameters
}
\description{
Return the number of parameters for the noise specification noiseSpec.
}
