% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{get_num_var_univariate_cont}
\alias{get_num_var_univariate_cont}
\title{Get the number of parameters for named variables of a univariate, continuous
model}
\usage{
get_num_var_univariate_cont(varName, modSpec)
}
\arguments{
\item{varName}{The variable name (either c or kappa)}

\item{modSpec}{The model specification}
}
\value{
The number of parameters
}
\description{
Get the number of parameters for named variables of a univariate, continuous
model
}
