% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{get_num_var_univariate_ord}
\alias{get_num_var_univariate_ord}
\title{Get the number of parameters for named variables of a univariate, ordinal 
model}
\usage{
get_num_var_univariate_ord(varName, modSpec)
}
\arguments{
\item{varName}{The variable name (b, tau, or beta)}

\item{modSpec}{The model specification}
}
\value{
The number of parameters
}
\description{
Get the number of parameters for named variables of a univariate, ordinal 
model
}
