% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{get_univariate_cont_transform_categories}
\alias{get_univariate_cont_transform_categories}
\title{For a given model specification of a univariate continuous variable, return a
vector that gives the categories used by param_constr2uncontr and
param_constr2unconstr.}
\usage{
get_univariate_cont_transform_categories(modSpec)
}
\arguments{
\item{modSpec}{The model specification}
}
\value{
A vector of categories
}
\description{
For a given model specification of a univariate continuous variable, return a
vector that gives the categories used by param_constr2uncontr and
param_constr2unconstr.
}
