% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{get_univariate_indices}
\alias{get_univariate_indices}
\title{For a multivariate model, get the indices in the full parameter vector
(th_y) correspdoning to a given univariate model. The correlation term is
ignored for conditionally dependent models. Exactly one of j, k, or i must
be specified (see the help for get_var_index_multivariate for more on j, k
and i).}
\usage{
get_univariate_indices(mod_spec, j = NA, k = NA, i = NA)
}
\arguments{
\item{mod_spec}{The model specification}

\item{j}{The ordinal index}

\item{k}{The continuous index}

\item{i}{The overall index}
}
\description{
For a multivariate model, get the indices in the full parameter vector
(th_y) correspdoning to a given univariate model. The correlation term is
ignored for conditionally dependent models. Exactly one of j, k, or i must
be specified (see the help for get_var_index_multivariate for more on j, k
and i).
}
