% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{get_univariate_variable_cases}
\alias{get_univariate_variable_cases}
\title{A helper function to check the files in data_dir to get the cases (fit
univariate models). Exactly one of j and k should be input.}
\usage{
get_univariate_variable_cases(data_dir, analysis_name, j = NA, k = NA)
}
\arguments{
\item{data_dir}{Directory where files are located}

\item{analysis_name}{analysis_name for file-searching}

\item{j}{Number of current ordinal variable}

\item{k}{Number of current continuous variable}
}
\description{
A helper function to check the files in data_dir to get the cases (fit
univariate models). Exactly one of j and k should be input.
}
