% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{get_var_index_multivariate}
\alias{get_var_index_multivariate}
\title{Description and stuff}
\usage{
get_var_index_multivariate(
  varName,
  modSpec,
  j = NA,
  k = NA,
  i = NA,
  i1 = NA,
  i2 = NA
)
}
\arguments{
\item{varName}{The variable name (a, tau, alpha, z)}

\item{modSpec}{The model specification}

\item{j}{The ordinal index}

\item{k}{The continuous index}

\item{i}{The overall index}

\item{i1}{The overall index for the first variable of a pair}

\item{i2}{The overall index for the second variable of a pair}
}
\value{
The number of parameters
}
\description{
Description and stuff
}
