% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{get_var_index_multivariate_mapping}
\alias{get_var_index_multivariate_mapping}
\title{A helper function to build the mapping for [get_var_index_multivariate_fast]}
\usage{
get_var_index_multivariate_mapping(mod_spec)
}
\arguments{
\item{mod_spec}{The model specification}
}
\value{
The mapping
}
\description{
Build the mapping between inputs and indices needed by the function
get_var_index_multivariate_fast.
}
