% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{get_z_full_fast}
\alias{get_z_full_fast}
\title{Extract the full correlation vector of length choose(J+K,2), accounting for
for the grouping structure specified by modSpec$cdepGroups.}
\usage{
get_z_full_fast(th_y, mapping, asMatrix = F)
}
\arguments{
\item{th_y}{The parameter vector}

\item{mapping}{The mapping object that supports quick indexing}

\item{asMatrix}{Whether or not to return a (J+K) by (J+K) matrix of correlations (default FALSE)}
}
\value{
The full correlation vector (or matrix)
}
\description{
Extract the full correlation vector of length choose(J+K,2), accounting for
for the grouping structure specified by modSpec$cdepGroups.
}
