% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntegInfo_theta_y.R
\name{getIntegInfo_theta_y}
\alias{getIntegInfo_theta_y}
\title{Get integration info for log-likelihood calculation for theta_y}
\usage{
getIntegInfo_theta_y(theta_y_list, Y)
}
\arguments{
\item{theta_y_list}{theta_y as a list}

\item{Y}{Matrix of dependent variable observations}
}
\value{
A list consisting of the matrix doIntegral and array intLim [see details]
}
\description{
\code{getIntegInfo_theta_y} determines the integration limits for the log-likelihood calculation for theta_y
}
\details{
Determine the integration limits. For ordinal variables, this comes from tau in theta. For missing data, the integration limit is -Inf to Inf. Two variables are returned in a list:

(a) doIntegral, which is a [(J+K) x N] matrix J+K is the number of variables and N is the number of observations. doIntegral is T if an integration must be done and F otherwise. F occurs for non-missing, continuous observations.

(b) intLim, which is a [(J+K x N x 2] array and provides the integration limits for the maximum likelihood calculation
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
