% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_theta_y.R
\name{init_theta_y}
\alias{init_theta_y}
\title{Initialize parameter vector for the mixed cumulative probit model}
\usage{
init_theta_y(x, Y, J, corr = T, hetero = T)
}
\arguments{
\item{x}{Vector of indepenent variable observations}

\item{Y}{Matrix of dependent variable observations}

\item{J}{Number of latent variables}

\item{corr}{Whether to include correlations across latent variables [default True]}

\item{heter0}{Whether to allow heteroskedasticity [default True]}
}
\value{
theta_y as a list
}
\description{
\code{init_theta_y} initializes the parameter vector, theta_y (e.g., for maximum likelihood estimation or adaptive Metropolis sampling). Individual variable fits are done by calling fitGenCrraOrd and fitGenCrra and all correlations are set to zero (off diagonal terms in Sigma).
}
\seealso{
simConstHaz, fitConstHaz, fitLinHaz
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
