% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{init_univariate_ord}
\alias{init_univariate_ord}
\title{Initialize the parameter vector for a univariate ordinal model}
\usage{
init_univariate_ord(x, v, modSpec)
}
\arguments{
\item{x}{The vector of independent variables}

\item{v}{The vector of ordinal responses}

\item{modSpec}{The model specification}
}
\value{
An initialization for the ordinal parameter vector, th_v0
}
\description{
Initialize the parameter vector for a univariate ordinal model
}
