% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{load_best_univariate_model}
\alias{load_best_univariate_model}
\title{Load the best univariate model using saved univariate cross-validation
results}
\usage{
load_best_univariate_model(
  data_dir,
  analysis_name,
  var_name,
  fold = NA,
  load_data = FALSE
)
}
\arguments{
\item{data_dir}{Directory to save univariate model parameters}

\item{analysis_name}{analysis_name for file-naming}

\item{var_name}{The variable name for the model to load}

\item{fold}{Fold number. If NA, pull from main solution.}

\item{load_data}{Whether to also load the data vector (x and y) (default:
FALSE)}
}
\value{
A list with the parameter vector (th_y), model specification
  (mod_spec). If load_data is TRUE, the data vectors (x and y) are also in
  the list.
}
\description{
The cross validation data file (cv_data) must already exists
in the analysis directory and the input variable name, var_name, must be
a variable in cv_data.
}
