% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{load_cat_list}
\alias{load_cat_list}
\title{Load a category list from file}
\usage{
load_cat_list(file_path)
}
\arguments{
\item{file_path}{The file path for the .txt file containing ordinal variable
category specifications}
}
\value{
A named list where the names are variable names and the elements are
  category spefications
}
\description{
Load a category list from input text file. The text file should alternate
lines with variable names and category specifications, e.g.:

var1
a c d b
var2
{b a} -1 {c d e} 4
...

See [parse_cat_spec]
yada
}
