% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{load_cp_data}
\alias{load_cp_data}
\title{Load cumulative probit data from file}
\usage{
load_cp_data(data_file, var_info)
}
\arguments{
\item{data_file}{The .csv data file}

\item{var_info}{The data frame with variable information (see
[load_var_info])}
}
\value{
A named list containing 1) a data frame containing the cumulative probit
data ("cp_df") and 2) the cumulative probit problem file ("problem")
}
\description{
Two inputs are needed, the path to the data file (data_file) and a data frame
with information on variables to use (var_info; see [load_var_info]).

Only variables with entries in var_info are kept (accounting for the
possibility that some are left/right variables).
}
\examples{
# Load data file path from package yada
data_file <- syanalysis_name.file("extdata","US.csv",package="yada")
print(data_file)

# Load var_info_file from package yada
var_info_file <- syanalysis_name.file("extdata","US_var_info.csv",package="yada")
var_info <- load_var_info(var_info_file)  # load var info file as data frame
head(var_info)

# Call load_cp_data to format data and problem file for cumulative probit
cp_data <- load_cp_data(data_file, var_info)
head(cp_data$cp_df)  # Top rows of cumulative probit data frame
head(cp_data$problem)  # cumulative probit problem as list

}
