% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_calculations.R
\name{mean_spec_int_to_str}
\alias{mean_spec_int_to_str}
\title{Convert from an integer to string representation of the mean specification}
\usage{
mean_spec_int_to_str(mean_spec_int)
}
\arguments{
\item{mean_spec_int}{The mean specification as an integer}
}
\value{
The mean specification as a string
}
\description{
Given an integer specification of the mean, convert to a string
specification. This is done for code readability. In the technical
documentation, the variable S^{(f)} gives the mean specifiation, but for
readability mean_spec is instead used in the code. An error is thrown for
invalid inputs. The valid cases are:

int	str
0	pow_law_ord
1	log_ord
2	lin_ord
3	pow_law
}
\examples{
mean_spec_int_to_str(0)
mean_spec_int_to_str(1)

}
