% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_functions.R
\name{meanSpec_int2str}
\alias{meanSpec_int2str}
\title{Given an integer specification of the mean, convert to a string
specification. This is done for code readability. In the technical
documentation, the variable S^{(f)} gives the mean specifiation, but for
readability meanSpec is instead used in the code. An error is thrown for
invalid inputs. The valid cases are:}
\usage{
meanSpec_int2str(meanSpec_int)
}
\arguments{
\item{meanSpec_int}{The mean specification as an integer}
}
\value{
meanSpec_str The mean specification as a string
}
\description{
int	str
0	powLawOrd
1	logwOrd
2	linOrd
3	powLaw
}
\examples{
meanSpec_int2str(0)
meanSpec_int2str(1)
}
