% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{merge_multiple_lr_var}
\alias{merge_multiple_lr_var}
\title{Merge multiple variables that have distinct left and right measurements}
\usage{
merge_multiple_lr_var(input_df, base_var, side_loc, side_labels, approach)
}
\arguments{
\item{input_df}{Original data frame to be manipulated.}

\item{base_var}{The base variables to be merged (a vector the same length as
approach)}

\item{side_loc}{Character string identifying if side labels are at the
"start" or "end" of the base variable name}

\item{side_labels}{Character string of left and right labels (in that order;
ex.: c('_L','_R')).}

\item{approach}{The merging approaches to take (a vector the same length as
baseVar)}
}
\value{
The merged data frame
}
\description{
The input is a data frame with multiple variables that have left and right
aspects (to merge a single variable, use merge_lr_var). Call merge_lr_var
for each variable to merge it. The input variables approach and base_var
should be vectors of the same length (the number of variables to merge).
sideLabels and sideLoc are as in [merge_lr_var], but are assumed to be the
same for all variables (consequently, it is not possible to mix prefixing,
"left_man_I1" and affixing, "man_I1_L"). Aside from these requirements,
approach, baseVar, sideLabels, and sideLoc behave identically in
[merge_multiple_lr_var] and [merge_lr_var].
}
\examples{
# Call merge_multiple_lr_var to combine left and right variables per
# user-specified approach
ex_df <- data.frame(ord_var_L = c(0,NA,1,NA,2,NA),
                    ord_var_R = c(2,1,0,NA,2,1),
                    cont_var_L = c(10.5,NA,14.0,NA,8.7,NA),
                    cont_var_R = c(12.5,3.2,7.0,NA,8.7,1.0))

merged_df <- merge_multiple_lr_var(input_df = ex_df,
                                   base_var = c("ord_var","cont_var"),
                                   side_loc = "end",
                                   side_labels = c("_L","_R"),
                                   approach = c("lowest","left"))
print(merged_df)

}
