% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_calculations.R
\name{noise_spec_int_to_str}
\alias{noise_spec_int_to_str}
\title{Convert from an integer to string representation of the noise specification}
\usage{
noise_spec_int_to_str(noise_spec_int)
}
\arguments{
\item{noise_spec_int}{The noise specification as an integer}
}
\value{
noise_spec_str The noise specification as a string
}
\description{
Given an integer specification of the noise, convert to a string
specification. This is done for code readability. In the technical
documentation, the variable S^{(sigma)} gives the noise specifiation, but
for readability noise_spec is instead used in the code. An error is thrown
for invalid inputs. The valid cases are:

int	str
0	const
1	lin_pos_int
}
\examples{
noise_spec_int_to_str(0)
noise_spec_int_to_str(1)

}
