% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_functions.R
\name{noiseSpec_int2str}
\alias{noiseSpec_int2str}
\title{Given an integer specification of the noise, convert to a string
specification. This is done for code readability. In the technical
documentation, the variable S^{(sigma)} gives the noise specifiation, but
for readability noiseSpec is instead used in the code. An error is thrown for
invalid inputs. The valid cases are:}
\usage{
noiseSpec_int2str(noiseSpec_int)
}
\arguments{
\item{noiseSpec_int}{The noise specification as an integer}
}
\value{
noiseSpec_str The noise specification as a string
}
\description{
int	str
0	const
1	lin_pos_int
}
\examples{
noiseSpec_int2str(0)
noiseSpec_int2str(1)
}
