% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_transforms.R
\name{param_constr2unconstr}
\alias{param_constr2unconstr}
\title{Transform a vector from a constrained to an unconstrained representation.
The categories are:
Category	Description
0		Unconstrained
1		Positive
2		-1 to 1
3		Positive difference}
\usage{
param_constr2unconstr(param, tfCatVect)
}
\arguments{
\item{param}{The vector of constrained parameters}

\item{tfCatVect}{The vector of transformation categories}
}
\value{
The vector of unconstrained parameters, param_bar
}
\description{
Transform a vector from a constrained to an unconstrained representation.
The categories are:
Category	Description
0		Unconstrained
1		Positive
2		-1 to 1
3		Positive difference
}
