% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_transforms.R
\name{param_unconstr2constr}
\alias{param_unconstr2constr}
\title{Transform a vector from an unconstrained to a constrained representation.
The categories are:
Category	Description
0		Unconstrained
1		Positive
2		-1 to 1
3		Positive difference}
\usage{
param_unconstr2constr(param_bar, tfCatVect)
}
\arguments{
\item{param_bar}{The vector of unconstrained parameters}

\item{tfCatVect}{The vector of transformation categories}
}
\value{
The vector of constrained parameters, param
}
\description{
Transform a vector from an unconstrained to a constrained representation.
The categories are:
Category	Description
0		Unconstrained
1		Positive
2		-1 to 1
3		Positive difference
}
