% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{parse_joined_model}
\alias{parse_joined_model}
\title{Parse a joined model to get the individual mean and noise specs}
\usage{
parse_joined_model(joined_model)
}
\arguments{
\item{joined_model}{The string representing the joined model.}
}
\value{
A vecto consisting of the mean spec and noise spec
}
\description{
The input is a string representing a mean/noise pair such as
"pow_law_ord_const". Return a vector in which the first element is the
mean spec and the second element the noise spec, c("pow_law_ord","const").
}
