% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{parse_NA_values}
\alias{parse_NA_values}
\title{Parse missing value information from var_info}
\usage{
parse_NA_values(missing_values)
}
\arguments{
\item{missing_values}{A string containing values that should be recoded as NA
separated by a comma}
}
\value{
A vector with the values that represent NA
}
\description{
The input is the Missing_Values column in var_info, which must have a length
equal to the number of columns in the data file. This function is for use in
[reformat_mcp_data] to ensure only values expected for each trait are available.
}
\examples{
# Call parse_NA_values to return a vector of individual proxies for NA
na_values <- parse_NA_values("-1,99,NA")
print(na_values)

}
