% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mixed_cum_probit_posterior.R
\name{plot_mixed_cum_probit_posterior}
\alias{plot_mixed_cum_probit_posterior}
\title{Plot posterior for mixed cumulative probit model}
\usage{
plot_mixed_cum_probit_posterior(theta_y_list, theta_x, xv, y, hp,
  knownX = NA)
}
\arguments{
\item{theta_y_list}{theta_y as a list}

\item{theta_x}{Parameterization for distribution of x (exponential rate)}

\item{xv}{A vector of ages at which to calculate posterior}

\item{y}{A new observation}

\item{hp}{Hyperparameters}

\item{knownX}{Known x value [optional]}
}
\value{
The vector of posterior probabilities
}
\description{
\code{plot_mixed_cum_probit_posterior} calculates and plots the posterior probability of an observation as a function of a vector xv, which is assumed to be uniformly sampled.
}
\seealso{
simConstHaz, fitConstHaz, fitLinHaz
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
