% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_x_posterior.R
\name{plot_x_posterior}
\alias{plot_x_posterior}
\title{Plot the result of a call to analyze_x_posterior}
\usage{
plot_x_posterior(post)
}
\arguments{
\item{post}{The output of analyze_x_posterior}
}
\description{
\code{plot_x_posterior} plots the result of a call to analyze_x_posterior. In particular, the density, median, mean, and 2.5% confidence interval locations are shown, as is the actual age (if known).
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
