% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powLaw.R
\name{powLaw}
\alias{powLaw}
\title{Power law with a constant offset}
\usage{
powLaw(x, th_w)
}
\arguments{
\item{x}{Vector of independent variable observations}

\item{th_w}{Vector of parameters with ordering [a,r,b,s,kap]}

\item{w}{Vector of dependent variable observations}

\item{a}{Multiplicative coefficient}

\item{r}{Scaling exponent}

\item{b}{Offset}

\item{s}{Baseline noise}

\item{kap}{Slope of noise (short for kappa) [Optional]}

\item{hetero}{Whether the model is heteroskedastic [Default FALSE]}
}
\description{
\code{powLaw} calculates the mean (h). \code{powLawSigma} calculates the noise (sigma, or sig for short). \code{powLawDensity} calculates the density. \code{powLawNegLogLik} calculates the negative log-likelihood. \code{powLawGradNegLogLik} calculates the gradient of the negative log-likelihood. \code{fitPowLaw} returns the maximum likelihood fit. \code{simPowLaw} creates simulated data.
}
\details{
We assume that the response variable w is distributed as

\deqn{w ~ N(h(x),sig(x)^2)}

where x is the independent variable, h(x) the mean, sig(x) the noise, and N denotes a normal distribution. If sig is independent of x, the model is homoskedastic. Otherwise, it is heteroskedastic. For an observation (w,x), the likelihood is

\deqn{l_w = 1/sqrt(2*pi)/sig*exp(-0.5*(w-h)^2/sig^2)}

The negative log-likelihood is

\deqn{eta_w = 0.5*log(2*pi) + log(sig) + 0.5*(w-h)^2/sig^2}

For the mean and noise, we adopt the parametric forms

\deqn{h = a*x^r + b}

and

\deqn{sig = s*(1 + kap*x)}

The components of the gradient of eta_w are

\deqn{eta_w_a   = -wbar / sig^2 * x^r                    }
\deqn{eta_w_r   = -wbar / sig^2 * x^r * log(x) * a       }
\deqn{eta_w_b   = -wbar / sig^2                          }
\deqn{eta_w_s   =  (1/sig - wbar^2/sig^3) * (1 + kap * x)}
\deqn{eta_w_kap =  (1/sig - wbar^2/sig^3) * s * x)       }

where wbar = w - h.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
