% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powLawMix.R
\name{powLawMixNegLogLik}
\alias{powLawMixNegLogLik}
\title{Power law with mixed ordinal and continuous observations}
\usage{
powLawMixNegLogLik(th_y, x, Y, hp, transformVar = F)
}
\arguments{
\item{th_y}{Parameter vector with ordering th_y = [rho,tau,a,r,b,s,kappa]}

\item{x}{Vector of independent variable observations}

\item{hp}{Hyperparameters that, among other things, specify how to unpack th_y}

\item{vstar}{Vector of latent dependent variable observations (for v^*)}

\item{v}{Vector of dependent variable observations (ordinal)}

\item{w}{Vector of dependent variable observations (continuous)}

\item{rho}{Scaling exponent (ordinal)}

\item{a}{Multiplicative coefficient (continuous)}

\item{r}{Scaling exponent (continuous)}

\item{b}{Offset (continuous)}

\item{s}{Baseline noise}

\item{kappa}{Slope of noise [Optional]}

\item{hetero}{Whether the model is heteroskedastic [Default FALSE]}
}
\description{
\code{powLawMixNegLogLik} calculates the negative log-likelihood. \code{powLawMixGradNegLogLik} calculates the gradient of the negative log-likelihood. \code{extract_th_v} extracts the parameterization for a single ordinal variable from the parameter vector th_y. \code{extract_th_w} extracts the parameterization for a single continuous variable from the parameter vector th_y.
}
\details{
We assume a mixture of J ordinal variables as described in yada::powLawOrd and K continuous variables as described in yada::powLaw. The calculation of the mean and scale term on the noise, s, is unchanged for the mixed case. However, the scaling term, kappa, is common across variables. The ordering of the complete parameter vector th_y is

\deqn{th_y = [rho,tau,a,r,b,s,kappa]}

The length of each vector in th_y is

Variable  Length
rho       J
tau       M1 + M2 + ... + Mj + ... + MJ, where Mj+1 is the number of ordinal
          categories for the j-th ordinal variable
a         K
r         K
b         K
s         J + K
kappa     1

The extension of the negative log-likelihod calculation from the single variables case ordinal/continuous case to the multi-variable mixed case is straightforward: each variable contributes independently to the sum. Extension of the gradient calculation is similarly straightforward: aside from kappa, which is shared across variables, the calculation is unchanged. For kappa, a sum across variables is needed.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
