% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processKidStatsInput.R
\name{processKidStatsInput}
\alias{processKidStatsInput}
\title{Do calculations for input KidStats data}
\usage{
processKidStatsInput(kidStatsInput, cumProbitModel, catList, xknown = NA)
}
\description{
Do calculations for input KidStats data
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
