% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{reformat_mcp_data}
\alias{reformat_mcp_data}
\title{Load cumulative probit data from file}
\usage{
reformat_mcp_data(data_file_path, var_info, save_file = F)
}
\arguments{
\item{data_file_path}{The .csv data file path}

\item{var_info}{The data frame with variable information (see
[load_var_info])}

\item{save_file}{Logical whether to save the resulting data frame as 
"*_reformatted.csv" in the data_file_path}
}
\value{
A data frame containing reformatted data for the mcp algorithm
}
\description{
Two inputs are needed, the path to the data file (data_file) and a data frame
with information on variables to use (var_info; see [load_var_info]).

Only variables with entries in var_info are kept (accounting for the
possibility that some are left/right variables).
}
