% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{remove_missing_variables}
\alias{remove_missing_variables}
\title{For a single observation, remove missing variables and perform a number of
related tasks (see Description)}
\usage{
remove_missing_variables(y0, mod_spec0)
}
\arguments{
\item{y0}{The unreduced response vector}

\item{mod_spec}{The unreduced model specification}
}
\value{
The list with y, mod_spec, mapping, mapping0, ind, and keep
}
\description{
For a single observation with response vector y0 and model specification
mod_spec0, modify mod_spec0 to account for missing observations in y0, which
are indicated with NA. In addition, create additional variables to support
calculation of the negative log-likelihood for the observation. (a) mapping:
a list of index mappings for the reduced data observation (see
get_var_index_multivariate_mapping and get_var_index_multivariate_fast).
(b) mapping0: the mapping for the unreduced problem. (c) ind: a vector that
subsets the full parameter vector to account for the reduction. (d) keep:
a boolean vector indicating which variables were kept in the reduction.
}
