% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{renumber_groups}
\alias{renumber_groups}
\title{Renumber the groups in the input vector group0 so that the group numbering is
sequential integers}
\usage{
renumber_groups(groups0)
}
\arguments{
\item{groups0}{The starting group vector, probably with gaps in it}
}
\value{
The updated group vector
}
\description{
Renumber the groups in the input vector group0 so that the group numbering
is sequential integers (this is typically done after subsetting an original
group vector).
}
