% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2Sigma.R
\name{s2Sigma}
\alias{s2Sigma}
\title{Convert covariance matrix from vector to matrix representation}
\usage{
s2Sigma(s)
}
\arguments{
\item{s}{The covariance matrix in vector form}

\item{Sigma}{The covariance matrix in matrix form}
}
\value{
The covariance matrix Sigma
}
\description{
\code{s2Sigma} converts the input vector s into an equivalent covariance matrix Sigma. The ordering of s is by row: Sigma_11, Sigma_12,...,Sigma_22,....
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
