% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{sample_x_posterior}
\alias{sample_x_posterior}
\title{Sample from the posterior density of x}
\usage{
sample_x_posterior(
  y,
  th_x,
  th_y,
  mod_spec,
  num_samp,
  thinning = 1,
  prop_rescale = 0.1
)
}
\arguments{
\item{y}{The response vector for a single observation}

\item{th_x}{Parameterization for prior on x}

\item{th_y}{Parameterization for likelihood}

\item{mod_spec}{The model specification}

\item{xcalc}{A vector of ages at which to calculate the posterior probability}
}
\value{
A vector of posterior probabilities
}
\description{
Do Metropolis-Hastings sampling of the posterior probability
p(x|y,theta_x,theta_y) for the mixed cumulative probit model given the
response vector y.
}
