% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_theta_y_sample.R
\name{save_theta_y_sample}
\alias{save_theta_y_sample}
\title{Save a sample of theta_y from the adaptive Metropolis algorithm}
\usage{
save_theta_y_sample(fileName, thetaList, logLikVect, x, Y, hp,
  varNames = NA, known = NA)
}
\arguments{
\item{fileName}{The save file}

\item{thetaList}{A list of samples of theta_y (samples are a list, not a vector)}

\item{x}{Vector of indepenent variable observations}

\item{Y}{Matrix of dependent variable observations}

\item{hp}{Hyperparameters}

\item{varNames}{Vector of variable names [optional]}

\item{known}{List object of known simulation values [optional]}
}
\description{
\code{save_theta_y_sample} saves to file the result of sampling theta_y using the Metropolis algorithm
}
\seealso{
simConstHaz, fitConstHaz, fitLinHaz
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
