% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigma2s.R
\name{Sigma2s}
\alias{Sigma2s}
\title{Convert covariance matrix from matrix to vector representation}
\usage{
Sigma2s(Sigma)
}
\arguments{
\item{Sigma}{The covariance matrix in matrix form}

\item{s}{The covariance matrix in vector form}
}
\value{
The vector s representing the covariance matrix
}
\description{
\code{Sigma2s} converts the input covariance matrix Sigma into an equivalent vector s. The ordering of s is by row: Sigma_11, Sigma_12,...,Sigma_22,....
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
