% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{hsiler}
\alias{hsiler}
\alias{chsiler}
\alias{ssiler}
\alias{dsiler}
\alias{psiler}
\alias{qsiler}
\alias{rsiler}
\alias{nllsiler}
\alias{hesiler}
\title{Siler hazard model}
\usage{
hsiler(x, a)

chsiler(x, a, x0 = 0)

ssiler(x, a, x0 = 0)

dsiler(x, a, x0 = 0)

psiler(x, a, x0 = 0)

qsiler(qvect, a, x0 = 0)

rsiler(N, a, x0 = 0)

nllsiler(a, x, x0 = 0)

hesiler(a, x, x0 = 0)
}
\arguments{
\item{x}{the ending point}

\item{a}{parameterization vector}

\item{x0}{the initial point}

\item{qvect}{vector of quantiles}

\item{N}{number of observations}
}
\value{
A numeric vector of length equal to the input
}
\description{
\code{dsiler} gives the probability density function, \code{psiler} the cumulative density function, \code{qsiler} the quantile function, \code{rsiler} random draws, \code{hsiler} the hazard, \code{chsiler} the cumulative hazard, \code{ssiler} the survival, \code{nllsiler} the negative log-likelihood, and \code{hesiler} the Hessian.
}
\details{
The Siler hazard is

\deqn{\lambda(x) = a_1 \exp(-a_2 \, x) + a_3 + a_4 \exp(a_5 \, x)}

where \eqn{a_i} (index notation) is the parameter vector. The cumulative hazard is found by integrating the hazard from some initial point \eqn{x_0} to \eqn{x},

\deqn{\Lambda(x_0,x) = -\frac{a_1}{a_2} (e^{-a_2 x} - e^{-a_2 x_0}) + a_3 (x-x_0) + \frac{a_4}{a_5} (e^{a_5 x} - e^{a_5 x_0})}.

If \eqn{x_0} (optional) is not input it is set equal to 0. The survival and cumulative density function are, respectively,

\deqn{S(x_0,x) = \exp(-\Lambda(x_0,x)) = \frac{S(0,x)}{S(0,x_0)} = \frac{S(x)}{S(x_0)}}

and

\deqn{F(x_0,x) = 1 - S(x_0,x)}.

The probability density function is

\deqn{f(x_0,x) = \frac{d}{dx} F(x_0,x) = \lambda(x) S(x_0,x)}.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
\keyword{Siler,}
\keyword{demography}
\keyword{hazard,}
\keyword{survival,}
