% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{sim_multivariate}
\alias{sim_multivariate}
\title{Create simulated observations for a multivarate mixed cumulative probit model}
\usage{
sim_multivariate(th_y, mod_spec, N = NA, th_x = NA, x = NA)
}
\arguments{
\item{th_y}{The parameter vector for the y-model to simulate with}

\item{mod_spec}{The model specification (for the y-model)}

\item{N}{The number of samples to simulate}

\item{th_x}{The parameterization for x}

\item{x}{The vector of independent variables}
}
\value{
A list object of simulated data containing x, Y, and Ystar
}
\description{
Create simulated observations for a multivariate mixed cumulative probit
model. For x, either the number of samples (N) and a parameter vector (th_x)
must be given, or the full vector (x) must be given.
}
