% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLinHaz.R
\name{simLinHaz}
\alias{simLinHaz}
\title{Simulate event history data with linear event and constant censoring hazards}
\usage{
simLinHaz(n, b1, a1, b2)
}
\arguments{
\item{n}{Number of observations to simulate}

\item{b1}{Intercept for the linear event hazard}

\item{a1}{Slope for the linear event hazard}

\item{b2}{Censoring hazard}
}
\value{
A Surv object (survival package) with:
\describe{
  \item{\code{time}}{The event or censoring time}
  \item{\code{status}}{The status indicator: TRUE for an event and FALSE for censoring}
}
}
\description{
\code{simLinHaz} generates simulated event history data assuming the event hazard is linear, hazEvent = b1 + a1*t, and the censoring hazard is constant, hazCens = b2.
}
\seealso{
simConstHaz, fitConstHaz, fitLinHaz
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
