% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMixedCumProbit.R
\name{simMixedCumProbit}
\alias{simMixedCumProbit}
\title{Created simulated data for a mixed cumulative probit analysis}
\usage{
simMixedCumProbit(theta_y_list, N, theta_x, hp, missDataProp = 0)
}
\arguments{
\item{theta_y_list}{theta_y as a list}

\item{N}{number of simulated observations}

\item{theta_x}{Rate parameter for independent variable draw from exponential distribution}

\item{hp}{Hyperparameters (needed to guide generation of simulated data)}

\item{missDataProp}{Proportion of missing data [default 0]}
}
\value{
sim A list of simulated data variables
}
\description{
\code{simMixedCumProbit} creates simulated data for a mixed cumulative probit analysis.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
