% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{solve_cont_problem}
\alias{solve_cont_problem}
\title{Solve a univariate continuous problem}
\usage{
solve_cont_problem(data_dir, analysis_name, cont_prob)
}
\arguments{
\item{data_dir}{Data directory directory where the solution will be saved}

\item{cont_prob}{The continuous problem to solve (likely generated by
[build_univariate_cont_problems]}
}
\value{
Whether or not the fit succeeded (TRUE or FALSE)
}
\description{
Solve the input unaviriate continuous problem by doing a maximum likelihood
fit to find the best parameter vector. A save file is added to the data
directory (data_dir) in .rds format.
}
