% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{solve_ord_problem}
\alias{solve_ord_problem}
\title{Solve a univariate ordinal problem}
\usage{
solve_ord_problem(data_dir, analysis_name, ord_prob, anneal_seed = NA)
}
\arguments{
\item{data_dir}{The data directory with problems and results}

\item{analysis_name}{A unique analysis name (for the input data directory)}

\item{ord_prob}{The ordinal problem to solve (likely generated by
[build_univariate_ord_problems]}

\item{anneal_seed}{An optional seed to pass to fit_univariate_ord to make
annealing reproducible (default: NA, not used)}
}
\value{
Whether or not the fit succeeded (TRUE or FALSE)
}
\description{
Solve the input univariate ordinal problem by doing a maximum likelihood
fit to find the best parameter vector. A save file is added to the data
directory (data_dir) in .rds format.
}
