% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvey_uncorr_homo.R
\name{solvey_uncorr_homo}
\alias{solvey_uncorr_homo}
\title{Do a maximum likelihood fit for theta_y for the uncorrelated, homoskedastic case}
\usage{
solvey_uncorr_homo(x, Y, hp, verbose = F)
}
\arguments{
\item{x}{A vector of independent variables [length N]}

\item{Y}{A matrix of ordinal and/or continuous response variables [(J+K) by N]}

\item{hp}{A list of hyperparameters}

\item{verbose}{Whether to print out information as the optimization proceeds (default FALSE)}
}
\value{
The fit as a list, theta_y_list
}
\description{
x is a vector of length N, Y a matrix of dimensions (J+K) by N,
and hp a list of hyperparameters. For each variable j (ordinal) and k
(continuous), do a maximum likelihood fit. These fits are entirely
independent for the uncorrelated, homoskedastic case.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
