% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_y_plot_corr.R
\name{theta_y_plot_corr}
\alias{theta_y_plot_corr}
\title{Plot the correlation vector in theta_y across Metropolis samples}
\usage{
theta_y_plot_corr(bundle, sub1 = NA, sub2 = NA, decim = 1,
  fileName = NA)
}
\arguments{
\item{bundle}{The bundle of variables (see \code{bundle_theta_y_sample})}

\item{decim}{An integer for sampling (decimating) data to reduce the number of points for very large samples [optional; default = 1 for all samples]}

\item{fileName}{Filename for pdf save file [optional]}

\item{bundle}{For optional subsetting of variables; first index [optional]}

\item{bundle}{For optional subsetting of variables; second index [optional]}
}
\description{
\code{theta_y_plot_corr} plots the correlation vector, s, in theta_y across Metropolis samples given a bundle of relevant variables generated by \code{bundle_theta_y_sample}.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
