% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_y_plot_tau.R
\name{theta_y_plot_tau}
\alias{theta_y_plot_tau}
\title{Plot tau in theta_y across Metropolis samples}
\usage{
theta_y_plot_tau(bundle, j, decim = 1, fileName = NA)
}
\arguments{
\item{bundle}{The bundle of variables (see \code{bundle_theta_y_sample})}

\item{j}{Index for which set of tau variables to plot}

\item{decim}{An integer for sampling (decimating) data to reduce the number of points for very large samples [optional; default = 1 for all samples]}

\item{fileName}{Filename for pdf save file [optional]}
}
\description{
\code{theta_y_plot_tau} plots tau in theta_y for a given ordinal index across Metropolis samples given a bundle of relevant variables generated by \code{bundle_theta_y_sample}.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
