% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_yIsValid.R
\name{theta_yIsValid}
\alias{theta_yIsValid}
\title{Check if theta_y is valid for sampling}
\usage{
theta_yIsValid(theta_y_list, hp, forceFinite = F, x = x, Y = Y)
}
\arguments{
\item{theta_y_list}{theta_y as a list}

\item{hp}{Hyperparameters}

\item{forceFinite}{Whether to ensure likelihood is finite [optional]}

\item{x}{Vector of indepenent variable observations [optional]}

\item{Y}{Matrix of dependent variable observations [optional]}
}
\value{
Boolean indicating whether theta_y is valid
}
\description{
\code{theta_yIsValid} checks if the input theta_y (as a list) is valid for sampling. In particular, ensure that (1) tau is ordered, (2) all correlations are between -1 and 1, (3) Sigma is positive definite, (4) and [optional] the likelihood is finite.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
