% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_yList2Vect.R
\name{theta_yList2Vect}
\alias{theta_yList2Vect}
\title{Convert theta_y from list to vector format}
\usage{
theta_yList2Vect(theta_y_list)
}
\arguments{
\item{theta_y_list}{theta_y as a list}

\item{theta_y}{theta_y as a vector}
}
\value{
theta_y as a vector
}
\description{
\code{theta_yList2Vect} converts the variable theta_y formatted as a list to a vector. The ordering of the vector is [rho, a, r, b, tau, s].
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
