% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_yVect2List.R
\name{theta_yVect2List}
\alias{theta_yVect2List}
\title{Convert theta_y from vector to list format}
\usage{
theta_yVect2List(theta_y, hp)
}
\arguments{
\item{theta_y}{theta_y as a vector}

\item{hp}{Hyperparameters (needed to guide conversion)}
}
\value{
theta_y as a list (theta_y_list)
}
\description{
\code{theta_yVect2List} converts the variable theta_y formatted as a vector to a list. The ordering of the vector is [rho, a, r, b, tau, s].
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
