% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{univariate_batch_calc}
\alias{univariate_batch_calc}
\title{Calculate point estimate and credible interval for test data
  (univariate)}
\usage{
univariate_batch_calc(
  data_dir,
  analysis_name,
  test_samp,
  demo_cols,
  input_cols,
  ci_type,
  th_x,
  xcalc,
  seed,
  save_file = T
)
}
\arguments{
\item{data_dir}{The directory from which models should be imported}

\item{analysis_name}{Unique identifier for the analysis}

\item{test_samp}{A data frame with test observations}

\item{demo_cols}{A vector representing demographic column numbers}

\item{input_cols}{A vector representing the input (predictor) columns}

\item{ci_type}{Whether the credible intervals should be calculated using 
"hdi" or "quantiles"}

\item{th_x}{The prior parameterization on x}

\item{xcalc}{A vector of evenly-spaced values at which to calculate the 
posterior density}

\item{seed}{A numeric value to set the seed for replication}

\item{save_file}{A logical for whether the results should be saved. 
(default: T)}
}
\description{
A wrapper function that calculates the point estimate and 
  credible interval for all observations in the provided test sample
  by input variable using univariate models. 
  The function prints the observation number currently being predicted and 
  also returns the top 6 rows of a completed prediction table.
}
