% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCov.R
\name{updateCov}
\alias{updateCov}
\title{Recursively update a covariance matrix}
\usage{
updateCov(X, covObj = NA)
}
\arguments{
\item{X}{A new measurement}

\item{covObj}{A list with the current covariance matrix, mean, and number of observations}
}
\value{
covObj, a list with the updated covariance matrix, mean, and number of observations
}
\description{
Recursively update a covariance matrix (and mean) by adding new observations.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
