% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{vis_cont_fit}
\alias{vis_cont_fit}
\title{Visualize a continuous fit}
\usage{
vis_cont_fit(x, w, th_w, mod_spec, xplot = c(), th_w0 = c(), show_unc = F, ...)
}
\arguments{
\item{x}{The vector of x values for the variable}

\item{w}{The vector of response values for the variable}

\item{th_w}{The fit parameter vector}

\item{mod_spec}{The model specification}

\item{xplot}{(optional) A vector at which to calculate the fit for the plot.
If not provided, a sequence of equally spaced values of length hundred
from min(x) to max(x) is used.}

\item{th_w0}{(optional) A baseline fit to add to the plot (likely the known
value of the parameter vector from a simulation)}

\item{show_unc}{(default: FALSE) Whether to plot the uncertainty for the fit}

\item{...}{Additional arguments to pass to the plot call for the x-w scatter
plot (e.g., to set the x- and y-labels).}
}
\description{
Visualize a continuous fit by plotting the observed points (x,w) and the fit.
}
