% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{vis_ord_fit}
\alias{vis_ord_fit}
\title{Visualize an ordinal fit}
\usage{
vis_ord_fit(
  x,
  v,
  th_v,
  mod_spec,
  bin_bounds = c(),
  xplot = c(),
  th_v0 = c(),
  ...
)
}
\arguments{
\item{x}{The vector of x values for the variable}

\item{v}{The vector of response values for the variable}

\item{th_v}{The fit parameter vector}

\item{mod_spec}{The model specification}

\item{xplot}{(optional) A vector at which to calculate the fit for the plot.
If not provided, a sequence of equally spaced values of length hundred
from min(x) to max(x) is used.}

\item{th_v0}{(optional) A baseline fit to add to the plot (likely the known
value of the parameter vector from a simulation)}

\item{...}{Additional arguments to pass to the plot call for the x-w scatter
plot (e.g., to set the x- and y-labels).}
}
\description{
Visualize an ordinal fit by binning based on the x-values, calculating the
probability of each m-value for each bin using the v-values, and plotting
the fits.
}
