% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{write_continuous_report}
\alias{write_continuous_report}
\title{Make an Rmarkdown document in the results folder for a continuous variable.}
\usage{
write_continuous_report(data_dir, analysis_name, k, line_width = NA)
}
\arguments{
\item{data_dir}{Directory where problem and CV files are located.}

\item{analysis_name}{analysis_name for file-naming}

\item{k}{Number of current continuous variable}

\item{line_width}{Line width of rMarkdown file. Default is NA.}
}
\description{
Make an Rmarkdown document in the results folder for a continuous variable.
}
