% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{write_matrix}
\alias{write_matrix}
\title{A helper function to write a matrix to an Rmd file}
\usage{
write_matrix(mat, rm_file)
}
\arguments{
\item{mat}{Matrix to be written}

\item{rm_file}{rMarkdown file name}
}
\description{
A helper function to write a matrix to an Rmd file
}
