% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{write_ordinal_report}
\alias{write_ordinal_report}
\title{Make an Rmarkdown document in the results folder for an ordinal variable.}
\usage{
write_ordinal_report(data_dir, analysis_name, j, line_width = NA)
}
\arguments{
\item{data_dir}{Directory where the problem and CV files are}

\item{analysis_name}{analysis_name for file-naming}

\item{j}{The ordinal variable number}

\item{line_width}{User-defined line widths. Default is NA}
}
\description{
Wrapper function that reads in univariate ordinal information
and generates an rMarkdown report in the same directory where the files are
housed. (Typically this is the results folder).
}
