% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yada_mcmc.R
\name{yada_sample_mcmc}
\alias{yada_sample_mcmc}
\title{Given an input negative log-likelihood and temperature, do MCMC sampling of
the parameter vector.}
\usage{
yada_sample_mcmc(neg_log_lik, theta0, numSamp, temp, propScale, ...)
}
\arguments{
\item{neg_log_lik}{The function for the negative log-likelihood}

\item{theta0}{The initial value of the parameter vector}

\item{numSamp}{The number of samples to make}

\item{temp}{The temperature}

\item{propScale}{A scalar value for the standard deviation of the proposal distribution}

\item{...}{Variables required by neg_log_lik}
}
\value{
A list with sampling information
}
\description{
Given an input negative log-likelihood and temperature, do MCMC sampling of
the parameter vector.
}
