% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yada_mcmc.R
\name{yada_tailored_optim}
\alias{yada_tailored_optim}
\title{A tailored optimization that works well for univariate models. While this
tailored optimzation usually rrequires more computations of the negative
log-likelihood function compared to other algorithms (e.g., directly calling
stats::optim with method='BFGS'), experience has shown that other algorithms
do not always find the minimum of the negative log-likelihood. Put slightly
differently, this tailored algorithm trades off number of computations for
improved robustness, which is likely almost always a trade-off the user
desires. The optimization consists of two steps. First, simulated annealing
is used, with an initial step at each temperature to set the standard
deviation of the (normal) proposal distribution to achieve an acceptance
ratio of about 0.23. Second, stats::optim is called using method='BFGS' to
refine the solution found by the simulated annealing.}
\usage{
yada_tailored_optim(neg_log_lik, theta0, ...)
}
\arguments{
\item{neg_log_lik}{The function for the negative log-likelihood}

\item{theta0}{The initial value of the parameter vector}

\item{...}{Variables required by neg_log_lik}
}
\value{
A list with sampling information and optimization results
}
\description{
A tailored optimization that works well for univariate models. While this
tailored optimzation usually rrequires more computations of the negative
log-likelihood function compared to other algorithms (e.g., directly calling
stats::optim with method='BFGS'), experience has shown that other algorithms
do not always find the minimum of the negative log-likelihood. Put slightly
differently, this tailored algorithm trades off number of computations for
improved robustness, which is likely almost always a trade-off the user
desires. The optimization consists of two steps. First, simulated annealing
is used, with an initial step at each temperature to set the standard
deviation of the (normal) proposal distribution to achieve an acceptance
ratio of about 0.23. Second, stats::optim is called using method='BFGS' to
refine the solution found by the simulated annealing.
}
