% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{build_cindep_model}
\alias{build_cindep_model}
\title{Build a multivariate, conditionally independent model using the univariate
cross-validation results.}
\usage{
build_cindep_model(data_dir, analysis_name, calc_se = FALSE)
}
\arguments{
\item{data_dir}{Directory to save univariate model parameters}

\item{analysis_name}{analysis_name for file-naming}

\item{calc_se}{Whether to calculate the standard error for th_y_bar}
}
\value{
A list with the parameter vector (th_y), model specification
  (mod_spec), and transformed parameter vector (th_y_bar). If calc_se is
  TRUE, the standard error for th_y_bar is added (th_y_bar_se).
}
\description{
Build a multivariate, conditionally independent model using the univariate
cross-validation results. Optionally, the standard error of the
unconstrained parameter vector (th_y_bar) is calculated to support the
re-scaling done in fit_multivariate.
}
