% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{calc_ci_ord}
\alias{calc_ci_ord}
\title{Calculate ordinal model confidence intervals}
\usage{
calc_ci_ord(ord_model, th_x, point_est, xcalc)
}
\arguments{
\item{ord_model}{List containing the parameter vector for y [th_y] and model
specification [mod_spec] for an ordinal univariate model (likely the 
output of `solve_ord_problem`)}

\item{th_x}{Parameterization for prior on x}

\item{point_est}{A character string with the type of point estimate to be 
returned: mean ("xmean"), median ("xmed"), or mode ("xmode")}

\item{xcalc}{A vector of evenly-spaced values at which to calculate the 
posterior density}
}
\description{
Wrapper function that takes an ordinal model and outputs the 
point estimate and 95% and 99% confidence intervals for each ordinal stage.
To allow reproducibility, an optional random number input seed seed can be
provided. The input seed should either be (1) NA [the default], (2) a single
integer, or (3) a vector of length M+1, where M is the number of ordinal
categories. If no seed is provided, one is drawn and treated as if it (a
single integer) was input. If a single integer is provided, it is used to
generate a vector of integers with  length M+1. The vector is used to
provide seeds for initializing the seed for each call to calc_x_posterior.
}
