% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{calc_cond_gauss_int_inputs}
\alias{calc_cond_gauss_int_inputs}
\title{Calculate the inputs to the conditional Gaussian integral}
\usage{
calc_cond_gauss_int_inputs(th_y, x, y, mapping)
}
\arguments{
\item{th_y}{The parameter vector}

\item{x}{The independent variable (a scalar)}

\item{y}{The response variables (a vector)}

\item{mapping}{The mapping for fast extraction of indices (see
[get_var_index_multivariate_mapping])}
}
\value{
A list with the mean vector (mean_vect), covariance matrix (cov_mat),
  low limit of integration (lo), high limit of integration (hi), and
  given/conditioned response variables (y_giv).
}
\description{
For a single observation, calculate the inputs to the conditional Gaussian
integral, which are mean_vect, coMat, lo, and hi. These are returned as a
list.
}
