% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{calc_joint}
\alias{calc_joint}
\title{Calculate the joint density probability p(x,y|th_x,th_y)}
\usage{
calc_joint(xcalc, y, th_x, th_y, mod_spec)
}
\arguments{
\item{xcalc}{A vector of ages at which to calculate the posterior probability}

\item{y}{The response vector for a single observation}

\item{th_x}{Parameterization for prior on x}

\item{th_y}{Parameterization for likelihood (conditional on x)}

\item{mod_spec}{The model specification}
}
\value{
A vector of joint probabilities
}
\description{
Calculate the joint density probability p(x,y|th_x,th_y) for the mixed
cumulative probit model at the points in xcalc. xcalc is assumed to be evenly
spaced.
}
