% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_calculations.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate the mean response for a yada mean specification}
\usage{
calc_mean(x, mean_spec, a = NA)
}
\arguments{
\item{x}{The independent variable}

\item{mean_spec}{The mean specification}

\item{a}{The parameter vector}
}
\value{
f(x,a), the mean (same length as x)
}
\description{
Calculate the mean, f, given x, the parameter vector a, and the mean
specification mean_spec.
}
