% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{calc_mean_univariate_ord}
\alias{calc_mean_univariate_ord}
\title{Calculate the mean for a univariate ordinal model}
\usage{
calc_mean_univariate_ord(x, th_v, mod_spec)
}
\arguments{
\item{x}{The vector of independent variables}

\item{th_v}{The ordinal parameter vector [b,tau,beta]}

\item{mod_spec}{The model specification}
}
\value{
The mean, g(x,b), which is the same length as x
}
\description{
Calcualte the mean, g, for a univariate ordinal model by extracting b from
th_v and calling calc_mean.
}
