% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{calc_neg_log_lik_cont}
\alias{calc_neg_log_lik_cont}
\title{Calcuate the negative log-likelihood for a continuous model}
\usage{
calc_neg_log_lik_cont(th_w, x, w, modSpec)
}
\arguments{
\item{x}{The vector of independent variables}

\item{w}{The vector of continuous responses}

\item{modSpec}{The model specification'}

\item{th_v}{The continuous parameter vector [c,kappa]}
}
\value{
The negative log-likelihood
}
\description{
Calcuate the negative log-likelihood for a continuous model
}
