% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{calc_neg_log_lik_multivariate}
\alias{calc_neg_log_lik_multivariate}
\title{Calculate the overall (summed) negative log-likelihood}
\usage{
calc_neg_log_lik_multivariate(th_y, calc_data, tf_cat_vect = NA)
}
\arguments{
\item{th_y}{The parameter vector}

\item{calc_data}{Data needed for a speedy negative log-likelihood calculation}

\item{tf_cat_vect}{A vector to transform th_y from an unconstrained to a
constrained representation (if necessary)}
}
\value{
The overall negative log-likelihood
}
\description{
Calculate the overall negative log-likelihood for a set of observations by
calling calc_neg_log_lik_vect_multivariate and summing the resulting vector.
}
