% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{calc_neg_log_lik_vect_cont}
\alias{calc_neg_log_lik_vect_cont}
\title{Calculate the vector of negative log-likelihoods for a continuous model}
\usage{
calc_neg_log_lik_vect_cont(th_w, x, w, mod_spec, tf_cat_vect = NA)
}
\arguments{
\item{x}{The vector of independent variables}

\item{w}{The vector of continuous responses}

\item{mod_spec}{The model specification}

\item{tf_cat_vect}{(Optional) The transformation category vector to transform
th_w to a constrained representation}

\item{th_v}{The continuous parameter vector [c,kappa]}
}
\value{
The vector of negative log-likelihoods, which is the same length as x
}
\description{
Calculate the vector of negative log-likelihoods for a continuous model
}
