% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{calc_neg_log_lik_vect_multivariate_chunk_outer}
\alias{calc_neg_log_lik_vect_multivariate_chunk_outer}
\title{The outer function for calculating the negative log-likelihood in chunks}
\usage{
calc_neg_log_lik_vect_multivariate_chunk_outer(
  th_y,
  calc_data,
  tf_cat_vect = NA,
  num_chunks = round(length(calc_data)/10)
)
}
\arguments{
\item{th_y}{The parameter vector}

\item{calc_data}{Data needed for a speedy negative log-likelihood calculation}

\item{tf_cat_vect}{A vector to transform th_y from an unconstrained to a
constrained representation (if necessary)}

\item{num_chunks}{The number of chunks to use (default:
round(length(calc_data)/10))}
}
\value{
The vector of negative log-likelihoods
}
\description{
Use a parallel for loop along with "chunks" of input data to calculate the
vector of negative log-likelihoods. For each chunk, a conventional for loop
is used for each negative log-likelihood calculation and a parallel for loop
is used to iterate across chunks. The output is identical to
calc_neg_log_lik_vect_multivariate, but chunking the data can lead to better
performance on some syanalysis_names.
}
