% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_calculations.R
\name{calc_noise}
\alias{calc_noise}
\title{Calculate the noise for a yada noise specification}
\usage{
calc_noise(x, noise_spec, alpha)
}
\arguments{
\item{x}{The independent variable}

\item{noise_spec}{The noise specification}

\item{alpha}{The parameter vector}

\item{The}{parameter vector}
}
\value{
sigma(x,alpha), the noise (same length as x)
}
\description{
Calculate the noise, sigma, given x, the parameter vector alpha, and the
noise specification noise_spec.
}
