% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{calc_noise_univariate_cont}
\alias{calc_noise_univariate_cont}
\title{Calculate the noise for a univariate continuous model}
\usage{
calc_noise_univariate_cont(x, th_w, mod_spec)
}
\arguments{
\item{x}{The vector of independent variables}

\item{th_w}{The continuous parameter vector [c,kappa]}

\item{mod_spec}{The model specification}
}
\value{
The noise, psi(x,kappa), which is the same length as x
}
\description{
Calculate the noise, psi, for a univariate continuous model by extracting
beta from th_w and calling calc_noise.
}
