% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_x_density.R
\name{calc_x_density}
\alias{calc_x_density}
\title{Calculate the density for x given parameterization theta_x}
\usage{
calc_x_density(x, theta_x)
}
\arguments{
\item{x}{A vector of ages at which to calculate the density}

\item{theta_x}{List with the fit type and parameter vector}
}
\value{
The vector of densities
}
\description{
\code{calc_x_density} calculates the density at x given a parameterization for the density of theta_x. Currently, the exponential distribution, Weibull mixtures, and uniform distribution are supported
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
