% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{calc_x_density}
\alias{calc_x_density}
\title{Calculate the density for a model th_x, p(x|th_x)}
\usage{
calc_x_density(x, th_x)
}
\arguments{
\item{x}{A vector of ages at which to calculate the density}

\item{th_x}{A list with the fit type (fit_type) and the fit parameter
specification (fit; see details)}
}
\description{
Calculate the density at x given a parameterization for the density of
th_x. Currently, the exponential distribution, Weibull mixtures, and
uniform distribution are supported. If fit_type is 'exponential' or
'uniform', the fit parameter (fit) is a vector, whereas for 'offset_weib_mix'
it is a list with the vectors lambda (weighting), shape, and scale.
}
