% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_x_posterior.R
\name{calc_x_posterior}
\alias{calc_x_posterior}
\title{Calculate the posterior probability for the mixed cumulative probit model}
\usage{
calc_x_posterior(xcalc, y, theta_x, theta_y, hp)
}
\arguments{
\item{xcalc}{A vector of ages at which to calculate the posterior probability}

\item{y}{An observation of traits used for the posterior calculation}

\item{theta_x}{Parameterization for prior on x}

\item{theta_y}{Parameterization for likelihood}

\item{hp}{Hyperparameters}
}
\value{
A vector of posterior probabilities
}
\description{
\code{calc_x_posterior} calculates the posterior probability (x|y,theta_x,theta_y) for the mixed cumulative probit model at the points in xcalc given the input parameterization y. xcalc is assumed to be evenly spaced.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
