% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{calc_x_posterior}
\alias{calc_x_posterior}
\title{Calculate the posterior density of x}
\usage{
calc_x_posterior(
  y,
  th_x,
  th_y,
  mod_spec,
  xcalc = c(),
  normalize = T,
  seed = NA
)
}
\arguments{
\item{y}{The response vector for a single observation}

\item{th_x}{Parameterization for prior on x}

\item{th_y}{Parameterization for likelihood}

\item{mod_spec}{The model specification}

\item{xcalc}{A vector of ages at which to calculate the posterior probability}

\item{normalize}{Whether or not to normalize to integrate to 1 (default:
TRUE).}

\item{seed}{An optional input seed to make sampling reproducibile
(default: NA, not used)}
}
\value{
A vector of posterior probabilities
}
\description{
Calculate the posterior probability p(x|y,theta_x,theta_y) for the mixed
cumulative probit model at the points in xcalc given the response vector
y. The calculation is for a single observation. If the normalization is done,
xcalc is assumed to be evenly spaced.
}
