% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinadics.R
\name{comb}
\alias{comb}
\title{Combinadic functions}
\usage{
comb(m, N, k)
}
\description{
Implementation of combinadic-related functions as described here:

James McCaffrey -- Generating the mth Lexicographical Element of a Mathematical Combination

Indexing begins from 0, not 1.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
