% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{fit_multivariate}
\alias{fit_multivariate}
\title{Fit a multivariate mixed cumulative probit model.}
\usage{
fit_multivariate(
  x,
  Y,
  mod_spec,
  cindep_model,
  prog_file = NA,
  save_file = NA,
  hjk_control = list()
)
}
\arguments{
\item{x}{The vector of independent variables}

\item{Y}{The matrix of responses}

\item{mod_spec}{The model specification}

\item{cindep_model}{A list specifying the conditionally independent model
(likely the output of build_cindep_model). cindep_model must contain
cindep_model$th_y_bar and cindep_model$th_y_bar_scale.}

\item{save_file}{String containing the progress file name}

\item{hjk_control}{A list passed to the function dfoptim::hjk()}
}
\value{
A list object containing  the best-fit parameter vector (th_y), and
  the unconstrained best-fit parameter (th_y_bar), the return object from
  the optimization (hjk_output), the model specification (mod_spec), and 
  a record of how many legs it took for optimization (legs)
}
\description{
Do a maximum likelihood fit of  multivariate mixed cumulative probit model
using the Hooke-Jeeves (hjk) algorithm. The parameter vector used in the
optimization has, roughly, a mean of zero and an expectation that the best
actual value is within about +/-1 of the starting value. This is
accomplished by subtracting the mean of the conditionally dependent fit and
dividing by the standard error of the conditionally dependent fit.
}
