% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{fit_univariate_cont}
\alias{fit_univariate_cont}
\title{Fit a univariate continuous model}
\usage{
fit_univariate_cont(x, w, mod_spec, req_conv = T)
}
\arguments{
\item{x}{The vector of independent variables}

\item{w}{The vector of continuous responses}

\item{mod_spec}{The model specification}

\item{req_conv}{Whether to require convergence of the optimization}
}
\value{
The continuous parameter vector, th_w
}
\description{
Fit a univariate continuous model
}
