% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{fit_univariate_ord}
\alias{fit_univariate_ord}
\title{Fit a univariate ordinal model}
\usage{
fit_univariate_ord(x, v, mod_spec, anneal = T, req_conv = T, anneal_seed = NA)
}
\arguments{
\item{x}{The vector of independent variables}

\item{v}{The vector of ordinal responses}

\item{mod_spec}{The model specification}

\item{anneal}{Whether or not to refine the initialization by annealing}

\item{req_conv}{Whether to require convergence of the optimization}

\item{anneal_seed}{An optional seed to pass to init_univariate_ord to make
annealing reproducible (default: NA, not used)}
}
\value{
The ordinal parameter vector, th_v
}
\description{
Fit a univariate ordinal model
}
