% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCrraOrd.R
\name{genCrraOrd}
\alias{genCrraOrd}
\title{Ordinal generalized CRRA parametric model}
\usage{
genCrraOrd(x, param, transformVar = F)
}
\arguments{
\item{x}{Vector of independent variable observations}

\item{y}{Vector of dependent variable observations}

\item{\deqn{\rho}}{Scaling exponent}

\item{\deqn{\sigma}}{Standard deviation of noise term [sig in code]}

\item{\deqn{\tau}}{Vector of category boundaries (one fewer elements than number of categories)}
}
\description{
\code{genCrraOrd} calculates the response (y). \code{genCrraOrdProb} calculates the category probabilities. fitGenCrraOrd returns the least squares fit. simGenCrraOrd creates simulated data.
}
\details{
The parametric form of the generalized Coefficient of Relative Risk Aversion (CRRA) for ordinal data is (no slope or intercept):

\deqn{y = x^(1-\rho)}

where \deqn{x} is the independent variable, \deqn{y} the latent dependent (response) variable, \deqn{\alpha} the multiplicative coefficient, \deqn{\rho} the scaling exponent, and  there is no offset.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
