% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R, R/data_io.R
\name{generate_ord_ci}
\alias{generate_ord_ci}
\title{Generate the confidence intervals for a given a cross-validated
ordinal variable}
\usage{
generate_ord_ci(
  data_dir,
  analysis_name,
  var_name,
  th_x,
  point_est,
  ci_type = "hdi",
  xcalc,
  save_file = F
)

generate_ord_ci(
  data_dir,
  analysis_name,
  var_name,
  th_x,
  point_est,
  ci_type = "hdi",
  xcalc,
  save_file = F
)
}
\arguments{
\item{data_dir}{The directory from which models should be imported}

\item{analysis_name}{Unique identifier for the analysis}

\item{var_name}{A character string with the variable of interest}

\item{th_x}{The prior parameterization on x}

\item{point_est}{A character string with the type of point estimate to be 
returned: mean ("xmean"), median ("xmed"), or mode ("xmode")}

\item{ci_type}{Whether the credible interval should be calculated using 
"quantiles" or highest density interval ("hdi"). (default is "hdi")}

\item{xcalc}{A vector of evenly-spaced values at which to calculate the 
posterior density}

\item{save_file}{A logical for whether the results should be saved. 
(default: T)}

\item{input_seed}{An optional seed that can be used to make results
reproducible. The input_seed must be either (1) NA / not provided (the
default), (2) a single integer, or (3) a vector with length M+1 (see
Description of calc_ci_ord for further details).}
}
\description{
This function returns the point estimate, 95% and 99%
confidence intervals for each ordinal stage of a given response variable.

Calculate a table providing the point estimate and 
  credible interval of a given univariate ordinal model
}
