% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{get_J}
\alias{get_J}
\title{Get J for a yada model specification}
\usage{
get_J(mod_spec)
}
\arguments{
\item{mod_spec}{The model specification}
}
\value{
The number of ordinal variables, J
}
\description{
Get the number of ordinal variables, J, given a model specification. If J is
not a field in mod_spec, 0 is returned.
}
