% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{get_K}
\alias{get_K}
\title{Get K for a yada model specification}
\usage{
get_K(mod_spec)
}
\arguments{
\item{mod_spec}{The model specification}
}
\value{
The number of continuous variables, K
}
\description{
Get the number of continuous variables, K, given a model specification. If K
is not a field in mod_spec, 0 is returned.
}
