% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_calculations.R
\name{get_mean_transform_categories}
\alias{get_mean_transform_categories}
\title{Return the transform categories (see param_constr_to_unconstr) for a yada mean
specification}
\usage{
get_mean_transform_categories(mean_spec)
}
\arguments{
\item{mean_spec}{The mean specification}
}
\value{
A vector of categories
}
\description{
For a given mean specification, return a vector that gives the categories
used by param_constr_to_unconstr and param_constr_to_unconstr.
}
