% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{get_multivariate_transform_categories}
\alias{get_multivariate_transform_categories}
\title{Return the categories used by [param_constr2uncontr] and
[param_constr_to_unconstr] for a multivariate yada model specification}
\usage{
get_multivariate_transform_categories(mod_spec)
}
\arguments{
\item{mod_spec}{The model specification}
}
\value{
A vector of categories
}
\description{
For the input model specification of a multivariate model, return a vector
that gives the categories used by [param_constr2uncontr] and
[param_constr_to_unconstr].
}
