% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_calculations.R
\name{get_noise_transform_categories}
\alias{get_noise_transform_categories}
\title{Return the transform categories (see param_constr_to_unconstr) for a yada noise
specification}
\usage{
get_noise_transform_categories(noise_spec)
}
\arguments{
\item{noise_spec}{The noise specification}
}
\value{
A vector of categories
}
\description{
For a given noise specification, return a vector that gives the categories
used by param_constr2uncontr and param_constr_to_unconstr.
}
