% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{get_non_singleton_groups}
\alias{get_non_singleton_groups}
\title{Get the non-singleton groups for the input group vector}
\usage{
get_non_singleton_groups(groupVect)
}
\arguments{
\item{groupVect}{The vector of group assignments}
}
\value{
A vector of non-singlteon groups
}
\description{
groupVect is a vector assigning each of its elements to a unique group.
Return the non-singleton groups in groupVect -- that is, the groups with more
than one member

For example:
groupVect <- c(1,2,NA,4,1,3,3)
print(get_non_singleton_groups(groupVect))
[1] 1 3
}
