% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_calculations.R
\name{get_num_var_mean}
\alias{get_num_var_mean}
\title{Get the number of variables for a yada mean specification}
\usage{
get_num_var_mean(mean_spec)
}
\arguments{
\item{mean_spec}{The mean specification}
}
\value{
The number of mean parameters
}
\description{
Return the number of parameters for the mean specification mean_spec.
}
