% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_calculations.R
\name{get_num_var_noise}
\alias{get_num_var_noise}
\title{Get the number of variables for a yada noise specification}
\usage{
get_num_var_noise(noise_spec)
}
\arguments{
\item{noise_spec}{The noise specification}
}
\value{
The number of noise parameters
}
\description{
Return the number of parameters for the noise specification noise_spec.
}
