% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{get_num_var_univariate_cont}
\alias{get_num_var_univariate_cont}
\title{Get the number of parameters for named variables of a univariate, continuous
model}
\usage{
get_num_var_univariate_cont(var_name, mod_spec)
}
\arguments{
\item{var_name}{The variable name (either c or kappa)}

\item{mod_spec}{The model specification}
}
\value{
The number of parameters
}
\description{
Get the number of parameters for named variables of a univariate, continuous
model (either c or kappa)
}
