% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{get_var_index_multivariate}
\alias{get_var_index_multivariate}
\title{Get the indices in the parameter vector th_y of named variables}
\usage{
get_var_index_multivariate(
  var_name,
  mod_spec,
  j = NA,
  k = NA,
  i = NA,
  i1 = NA,
  i2 = NA
)
}
\arguments{
\item{var_name}{The variable name (a, tau, alpha, z)}

\item{mod_spec}{The model specification}

\item{j}{The ordinal index}

\item{k}{The continuous index}

\item{i}{The overall index}

\item{i1}{The overall index for the first variable of a pair}

\item{i2}{The overall index for the second variable of a pair}
}
\value{
The indices in the full parameter vector, th_y
}
\description{
Get the indices in the full parameter vector, th_y, of a named variable.
Optionally, the variable index (j, k, or i) can be specified, or (for the
correlation, z, the variable indices i1 and i2. The overall index, i, equals
j for ordinal variables and J + k for continuous variables, where J is the
number of ordinal variables. If no index is specified, the full range of
indices for the input variable is returned. If the variable index (or
indices) is specified, the relevant subrange is returned. The following
table lists the five valid input patterns, where 1 indicates the index
variable is specified and a 0 indicates it is not (and is thus equal to the
default value, NA).
j   k   i  i1  i2
0   0   0   0   0    no index
1   0   0   0   0    j specified
0   1   0   0   0    k specified
0   0   1   0   0    i specified
0   0   0   1   1    i1 and i2 specified
}
