% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{get_var_index_multivariate_fast}
\alias{get_var_index_multivariate_fast}
\title{Quickly get the indices in the parameter vector th_y of named variables}
\usage{
get_var_index_multivariate_fast(
  var_name,
  mapping,
  j = NA,
  k = NA,
  i = NA,
  i1 = NA,
  i2 = NA
)
}
\arguments{
\item{var_name}{The variable name (a, tau, alpha, z)}

\item{j}{The ordinal index}

\item{k}{The continuous index}

\item{i}{The overall index}

\item{i1}{The overall index for the first variable of a pair}

\item{i2}{The overall index for the second variable of a pair}
}
\value{
The indices in the full parameter vector, th_y
}
\description{
Accomplishes the same task as [get_var_index_multivariate], but using a
pre-built mapping to increase the speed of the lookup.
}
