% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{get_var_index_univariate_ord}
\alias{get_var_index_univariate_ord}
\title{Get the variable indices for a univariate, ordinal model given the variable
name}
\usage{
get_var_index_univariate_ord(var_name, mod_spec)
}
\arguments{
\item{var_name}{The variable name (b, tau, or beta)}

\item{mod_spec}{The model specification}
}
\value{
The variable indices
}
\description{
Get the variable indices for a univariate, ordinal model given the variable
name (b, tau, or beta)
}
