% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{get_z_length}
\alias{get_z_length}
\title{Get the number of correlation terms for a yada model specification}
\usage{
get_z_length(mod_spec)
}
\arguments{
\item{mod_spec}{The model specification}
}
\value{
The number of correlation terms
}
\description{
Return the number of correlation terms (length of z)
}
