% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_models.R
\name{init_univariate_ord}
\alias{init_univariate_ord}
\title{Initialize a univariate ordinal model}
\usage{
init_univariate_ord(x, v, mod_spec, anneal = T, anneal_seed = NA)
}
\arguments{
\item{x}{The vector of independent variables}

\item{v}{The vector of ordinal responses}

\item{mod_spec}{The model specification}

\item{anneal}{Whether or not to refine the initialization by annealing}

\item{anneal_seed}{An optional seed to make annealing reproducible
(default: NA, not used)}
}
\value{
An initialization for the ordinal parameter vector, th_v0
}
\description{
Initialize the parameter vector for a univariate ordinal model
}
