% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_models.R
\name{is_cdep}
\alias{is_cdep}
\title{Is this model conditionally dependent?}
\usage{
is_cdep(mod_spec)
}
\arguments{
\item{mod_spec}{The model specification}
}
\value{
Whether or not the model is conditionally dependent
}
\description{
Return True for conditionally dependent models and False for conditionally
independent models. For the special case of a univariate model, False is
returned. Aside from this, the model is conditionally indepenendent if
(a) mod_spec$cdep_spec is 'indep' or (b) mod_spec$cdep_spec is 'dep', but
cdep_groups is all NA.
}
