% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_transforms.R
\name{param_constr2unconstr}
\alias{param_constr2unconstr}
\title{Transform a vector from a constrained to an unconstrained representation.
The categories and bounds are:
category	 min	max
0		-Inf	Inf
1		   0	Inf
2		  -1	 +1}
\usage{
param_constr2unconstr(param, transfMat)
}
\arguments{
\item{param}{The vector of constrained parameters}

\item{transfMat}{The matrix specifying the transform}
}
\value{
The vector unconstrained parameters, param_bar
}
\description{
Transform a vector from a constrained to an unconstrained representation.
The categories and bounds are:
category	 min	max
0		-Inf	Inf
1		   0	Inf
2		  -1	 +1
}
