% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_transforms.R
\name{param_unconstr2constr}
\alias{param_unconstr2constr}
\title{Transform a vector from an unconstrained to a unconstrained representation.
The categories and bounds are:
category	 min	max
0		-Inf	Inf
1		   0	Inf
2		  -1	 +1}
\usage{
param_unconstr2constr(param_bar, transfMat)
}
\arguments{
\item{param_bar}{The vector of unconstrained parameters}

\item{transfMat}{The matrix specifying the transform}
}
\value{
The vector of constrained parameters, param
}
\description{
Transform a vector from an unconstrained to a unconstrained representation.
The categories and bounds are:
category	 min	max
0		-Inf	Inf
1		   0	Inf
2		  -1	 +1
}
