% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_x_posterior}
\alias{plot_x_posterior}
\title{Plot the posterior density using the result of a call to analyze_x_posterior}
\usage{
plot_x_posterior(analysis, ...)
}
\arguments{
\item{analysis}{The result of a call to analyze_x_posterior}

\item{...}{Additional parameters for the line plot (e.g., the x- and
y-labels)}
}
\description{
x_post_obj is the result of a call to analyze_x_posterior. Plot the density,
including marking the mean and 95% confidence intervals. If the known age was
provided in the call to analyze_x_posterior, add it to the plot.
}
