% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powLaw.R
\name{powLaw}
\alias{powLaw}
\title{Power law with a constant offset}
\usage{
powLaw(x, th_w, transformVar = F)
}
\arguments{
\item{x}{Vector of independent variable observations}

\item{th_w}{Vector of parameters with ordering [a,r,b,s,kappa]}

\item{transformVar}{Whether a transformation of the parameterization is needed [Default FALSE]}

\item{w}{Vector of dependent variable observations}

\item{a}{Multiplicative coefficient}

\item{r}{Scaling exponent}

\item{b}{Offset}

\item{s}{Baseline noise}

\item{kappa}{Slope of noise [Optional]}

\item{hetSpec}{Specification for the heteroskedasticity [Default 'none']}
}
\description{
\code{powLaw} calculates the mean (h). \code{powLawSigma} calculates the noise (sigma, or sig for short). \code{powLawDensity} calculates the density. \code{powLawNegLogLikVect} calculates a vector of negative log-likelihood. \code{powLawNegLogLik} calculates the negative log-likelihood (sum of \code{powLawNegLogLikVect}). \code{fitPowLaw} returns the maximum likelihood fit. \code{simPowLaw} creates simulated data.
}
\details{
We assume that the response variable w is distributed as

\deqn{w ~ N(h(x),sig(x)^2)}

where x is the independent variable, h(x) the mean, sig(x) the noise, and N denotes a normal distribution. If sig is independent of x, the model is homoskedastic. Otherwise, it is heteroskedastic. For an observation (w,x), the likelihood is

\deqn{l_w = 1/sqrt(2*pi)/sig*exp(-0.5*(w-h)^2/sig^2)}

The negative log-likelihood is

\deqn{eta_w = 0.5*log(2*pi) + log(sig) + 0.5*(w-h)^2/sig^2}

For the mean and noise, we adopt the parametric forms

\deqn{h = a*x^r + b}

and

\deqn{sig = s}                    [hetSepc = 'none']
\deqn{sig = s*(1 + kappa*x)}      [hetSpec = 'sd_x']
\deqn{sig = s*(1 + kappa*a*x^r)}  [hetSpec = 'sd_resp']
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
