% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simConstHaz.R
\name{simConstHaz}
\alias{simConstHaz}
\title{Simulate event history data with constant event and censoring hazards}
\usage{
simConstHaz(n, a1, a0)
}
\arguments{
\item{n}{Number of observations to simulate}

\item{a1}{Event hazard}

\item{a0}{Censoring hazard}
}
\value{
A Surv object (survival package) with:
\describe{
  \item{\code{time}}{The event or censoring time}
  \item{\code{status}}{The status indicator: TRUE for an event and FALSE for censoring}
}
}
\description{
\code{simConstHaz} generates simulated event history data assuming the event hazard, a1, and censoring hazard, a0, are constants
}
\seealso{
simLinHaz, fitConstHaz, fitLinHaz
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
