% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yada_mcmc.R
\name{yada_sample_mcmc}
\alias{yada_sample_mcmc}
\title{Sample a likelihood function at a given temperature using an input negative
log-likelihood}
\usage{
yada_sample_mcmc(neg_log_lik, theta0, num_samp, temp, prop_scale, ...)
}
\arguments{
\item{neg_log_lik}{The function for the negative log-likelihood}

\item{theta0}{The initial value of the parameter vector}

\item{num_samp}{The number of samples to make}

\item{temp}{The temperature}

\item{prop_scale}{A scalar value for the standard deviation of the proposal
distribution}

\item{...}{Variables required by neg_log_lik}
}
\value{
A list containing starting and sampling information (see description)
}
\description{
Given an input negative log-likelihood and temperature, do MCMC sampling of
the parameter vector. Let eta be the value of the negative log-likelihood.
The function that is sampled is exp(-eta/temp), where temp is the
temperature; that is, the likelihood^(1/temp) is sampled. The proposal
is made using an independent normal draw for each variable in the parameter
vector, theta, with the standard deviation of the proposal distribution set
by the scalar parameter prop_scale (the same scale is used for all variables).

The output is a list with starting and sampling information ,and includes the
following named elements:

\itemize{
  \item{\code{theta0}}
    {The initial value of the parameter vector}
  \item{\code{eta0}}
    {The initial value of the negative log-likelihood}
  \item{\code{temp}}
    {The temperature}
  \item{\code{prop_scale}}
    {The scale (standard deviation) for the normal proposal distribution}
  \item{\code{theta_best}}
    {The best parameter vector found (lowest negative log-likelihood)}
  \item{\code{eta_best}}
    {The best (lowest) negative log-likelihood encountered}
  \item{\code{accept_vect}}
    {A boolean vector indicating which samples were accepted}
  \item{\code{eta_vect}}
    {A vector of negative log-likelihood values (for kept samples)}
  \item{\code{eta_vect}}
    {A matrix of sampled parameter vectors, size num_samp by length(theta)}
}
}
